Add axis to encoding

Add axis parameters to an encoding

vl_axis_x(spec, bandPosition = NULL, domain = NULL,
  domainColor = NULL, domainDash = NULL, domainDashOffset = NULL,
  domainOpacity = NULL, domainWidth = NULL, format = NULL,
  formatType = NULL, grid = NULL, gridColor = NULL,
  gridDash = NULL, gridDashOffset = NULL, gridOpacity = NULL,
  gridWidth = NULL, labelAlign = NULL, labelAngle = NULL,
  labelBaseline = NULL, labelBound = NULL, labelColor = NULL,
  labelFlush = NULL, labelFlushOffset = NULL, labelFont = NULL,
  labelFontSize = NULL, labelFontStyle = NULL,
  labelFontWeight = NULL, labelLimit = NULL, labelOpacity = NULL,
  labelOverlap = NULL, labelPadding = NULL, labelSeparation = NULL,
  labels = NULL, maxExtent = NULL, minExtent = NULL, offset = NULL,
  orient = NULL, position = NULL, tickColor = NULL,
  tickCount = NULL, tickDash = NULL, tickDashOffset = NULL,
  tickExtra = NULL, tickMinStep = NULL, tickOffset = NULL,
  tickOpacity = NULL, tickRound = NULL, tickSize = NULL,
  tickWidth = NULL, ticks = NULL, title = NULL, titleAlign = NULL,
  titleAnchor = NULL, titleAngle = NULL, titleBaseline = NULL,
  titleColor = NULL, titleFont = NULL, titleFontSize = NULL,
  titleFontStyle = NULL, titleFontWeight = NULL, titleLimit = NULL,
  titleOpacity = NULL, titlePadding = NULL, titleX = NULL,
  titleY = NULL, values = NULL, zindex = NULL)

vl_axis_y(spec, bandPosition = NULL, domain = NULL,
  domainColor = NULL, domainDash = NULL, domainDashOffset = NULL,
  domainOpacity = NULL, domainWidth = NULL, format = NULL,
  formatType = NULL, grid = NULL, gridColor = NULL,
  gridDash = NULL, gridDashOffset = NULL, gridOpacity = NULL,
  gridWidth = NULL, labelAlign = NULL, labelAngle = NULL,
  labelBaseline = NULL, labelBound = NULL, labelColor = NULL,
  labelFlush = NULL, labelFlushOffset = NULL, labelFont = NULL,
  labelFontSize = NULL, labelFontStyle = NULL,
  labelFontWeight = NULL, labelLimit = NULL, labelOpacity = NULL,
  labelOverlap = NULL, labelPadding = NULL, labelSeparation = NULL,
  labels = NULL, maxExtent = NULL, minExtent = NULL, offset = NULL,
  orient = NULL, position = NULL, tickColor = NULL,
  tickCount = NULL, tickDash = NULL, tickDashOffset = NULL,
  tickExtra = NULL, tickMinStep = NULL, tickOffset = NULL,
  tickOpacity = NULL, tickRound = NULL, tickSize = NULL,
  tickWidth = NULL, ticks = NULL, title = NULL, titleAlign = NULL,
  titleAnchor = NULL, titleAngle = NULL, titleBaseline = NULL,
  titleColor = NULL, titleFont = NULL, titleFontSize = NULL,
  titleFontStyle = NULL, titleFontWeight = NULL, titleLimit = NULL,
  titleOpacity = NULL, titlePadding = NULL, titleX = NULL,
  titleY = NULL, values = NULL, zindex = NULL)

vl_remove_axis_x(spec)

vl_remove_axis_y(spec)

Arguments

spec
An input vega-lite spec

bandPosition
(Axis) An interpolation fraction indicating where, for band scales, axis ticks should be positioned. A value of 0 places ticks at the left edge of their bands. A value of 0.5 places ticks in the middle of their bands.

Default value: 0.5

domain
(Axis) A boolean flag indicating if the domain (the axis baseline) should be included as part of the axis.

Default value: true

domainColor
(Axis) Color of axis domain line.

Default value: "gray".

domainDash
(Axis) An array of alternating [stroke, space] lengths for dashed domain lines.

domainDashOffset
(Axis) The pixel offset at which to start drawing with the domain dash array.

domainOpacity
(Axis) Opacity of the axis domain line.

domainWidth
(Axis) Stroke width of axis domain line

Default value: 1

format
(Axis) The text formatting pattern for labels of guides (axes, legends, headers) and text marks.

If the format type is "number" (e.g., for quantitative fields), this is D3's number format pattern.

If the format type is "time" (e.g., for temporal fields), this is D3's time format pattern.

See the format documentation for more examples.

Default value: Derived from numberFormat config for number format and from timeFormat config for time format.

formatType
(Axis) The format type for labels ("number" or "time").

Default value:

"time" for temporal fields and ordinal and nomimal fields with timeUnit.

"number" for quantitative fields as well as ordinal and nomimal fields without timeUnit.

grid
(Axis) A boolean flag indicating if grid lines should be included as part of the axis

Default value: true for continuous scales that are not binned; otherwise, false.

gridColor
(Axis) Color of gridlines.

Default value: "lightGray".

gridDash
(Axis) An array of alternating [stroke, space] lengths for dashed grid lines.

gridDashOffset
(Axis) The pixel offset at which to start drawing with the grid dash array.

gridOpacity
(Axis) The stroke opacity of grid (value between [0,1])

Default value: 1

gridWidth
(Axis) The grid width, in pixels.

Default value: 1

labelAlign
(Axis) Horizontal text alignment of axis tick labels, overriding the default setting for the current axis orientation.

labelAngle
(Axis) The rotation angle of the axis labels.

Default value: -90 for nominal and ordinal fields; 0 otherwise.

labelBaseline
(Axis) Vertical text baseline of axis tick labels, overriding the default setting for the current axis orientation. Can be "top", "middle", "bottom", or "alphabetic".

labelBound
(Axis) Indicates if labels should be hidden if they exceed the axis range. If false (the default) no bounds overlap analysis is performed. If true, labels will be hidden if they exceed the axis range by more than 1 pixel. If this property is a number, it specifies the pixel tolerance: the maximum amount by which a label bounding box may exceed the axis range.

Default value: false.

labelColor
(Axis) The color of the tick label, can be in hex color code or regular color name.

labelFlush
(Axis) Indicates if the first and last axis labels should be aligned flush with the scale range. Flush alignment for a horizontal axis will left-align the first label and right-align the last label. For vertical axes, bottom and top text baselines are applied instead. If this property is a number, it also indicates the number of pixels by which to offset the first and last labels; for example, a value of 2 will flush-align the first and last labels and also push them 2 pixels outward from the center of the axis. The additional adjustment can sometimes help the labels better visually group with corresponding axis ticks.

Default value: true for axis of a continuous x-scale. Otherwise, false.

labelFlushOffset
(Axis) Indicates the number of pixels by which to offset flush-adjusted labels. For example, a value of 2 will push flush-adjusted labels 2 pixels outward from the center of the axis. Offsets can help the labels better visually group with corresponding axis ticks.

Default value: 0.

labelFont
(Axis) The font of the tick label.

labelFontSize
(Axis) The font size of the label, in pixels.

labelFontStyle
(Axis) Font style of the title.

labelFontWeight
(Axis) Font weight of axis tick labels.

labelLimit
(Axis) Maximum allowed pixel width of axis tick labels.

Default value: 180

labelOpacity
(Axis) The opacity of the labels.

labelOverlap
(Axis) The strategy to use for resolving overlap of axis labels. If false (the default), no overlap reduction is attempted. If set to true or "parity", a strategy of removing every other label is used (this works well for standard linear axes). If set to "greedy", a linear scan of the labels is performed, removing any labels that overlaps with the last visible label (this often works better for log-scaled axes).

Default value: true for non-nominal fields with non-log scales; "greedy" for log scales; otherwise false.

labelPadding
(Axis) The padding, in pixels, between axis and text labels.

Default value: 2

labelSeparation
(Axis) The minimum separation that must be between label bounding boxes for them to be considered non-overlapping (default 0). This property is ignored if labelOverlap resolution is not enabled.

labels
(Axis) A boolean flag indicating if labels should be included as part of the axis.

Default value: true.

maxExtent
(Axis) The maximum extent in pixels that axis ticks and labels should use. This determines a maximum offset value for axis titles.

Default value: undefined.

minExtent
(Axis) The minimum extent in pixels that axis ticks and labels should use. This determines a minimum offset value for axis titles.

Default value: 30 for y-axis; undefined for x-axis.

offset
(Axis) The offset, in pixels, by which to displace the axis from the edge of the enclosing group or data rectangle.

Default value: derived from the axis config's offset (0 by default)

orient
(Axis) The orientation of the axis. One of "top", "bottom", "left" or "right". The orientation can be used to further specialize the axis type (e.g., a y-axis oriented towards the right edge of the chart).

Default value: "bottom" for x-axes and "left" for y-axes.

position
(Axis) The anchor position of the axis in pixels. For x-axes with top or bottom orientation, this sets the axis group x coordinate. For y-axes with left or right orientation, this sets the axis group y coordinate.

Default value: 0

tickColor
(Axis) The color of the axis's tick.

Default value: "gray"

tickCount
(Axis) A desired number of ticks, for axes visualizing quantitative scales. The resulting number may be different so that values are "nice" (multiples of 2, 5, 10) and lie within the underlying scale's range.

tickDash
(Axis) An array of alternating [stroke, space] lengths for dashed tick mark lines.

tickDashOffset
(Axis) The pixel offset at which to start drawing with the tick mark dash array.

tickExtra
(Axis) Boolean flag indicating if an extra axis tick should be added for the initial position of the axis. This flag is useful for styling axes for band scales such that ticks are placed on band boundaries rather in the middle of a band. Use in conjunction with "bandPosition": 1 and an axis "padding" value of 0.

tickMinStep
(Axis) The minimum desired step between axis ticks, in terms of scale domain values. For example, a value of 1 indicates that ticks should not be less than 1 unit apart. If tickMinStep is specified, the tickCount value will be adjusted, if necessary, to enforce the minimum step value.

Default value: undefined

tickOffset
(Axis) Position offset in pixels to apply to ticks, labels, and gridlines.

tickOpacity
(Axis) Opacity of the ticks.

tickRound
(Axis) Boolean flag indicating if pixel position values should be rounded to the nearest integer.

Default value: true

tickSize
(Axis) The size in pixels of axis ticks.

Default value: 5

tickWidth
(Axis) The width, in pixels, of ticks.

Default value: 1

ticks
(Axis) Boolean value that determines whether the axis should include ticks.

Default value: true

title
(Axis) A title for the field. If null, the title will be removed.

Default value: derived from the field's name and transformation function (aggregate, bin and timeUnit). If the field has an aggregate function, the function is displayed as part of the title (e.g., "Sum of Profit"). If the field is binned or has a time unit applied, the applied function is shown in parentheses (e.g., "Profit (binned)", "Transaction Date (year-month)"). Otherwise, the title is simply the field name.

Notes:

You can customize the default field title format by providing the fieldTitle property in the config or fieldTitle function via the compile function's options.

If both field definition's title and axis, header, or legend title are defined, axis/header/legend title will be used.

titleAlign
(Axis) Horizontal text alignment of axis titles.

titleAnchor
(Axis) Text anchor position for placing axis titles.

titleAngle
(Axis) Angle in degrees of axis titles.

titleBaseline
(Axis) Vertical text baseline for axis titles.

titleColor
(Axis) Color of the title, can be in hex color code or regular color name.

titleFont
(Axis) Font of the title. (e.g., "Helvetica Neue").

titleFontSize
(Axis) Font size of the title.

titleFontStyle
(Axis) Font style of the title.

titleFontWeight
(Axis) Font weight of the title. This can be either a string (e.g "bold", "normal") or a number (100, 200, 300, ..., 900 where "normal" = 400 and "bold" = 700).

titleLimit
(Axis) Maximum allowed pixel width of axis titles.

titleOpacity
(Axis) Opacity of the axis title.

titlePadding
(Axis) The padding, in pixels, between title and axis.

titleX
(Axis) X-coordinate of the axis title relative to the axis group.

titleY
(Axis) Y-coordinate of the axis title relative to the axis group.

values
(Axis) Explicitly set the visible axis tick values.

zindex
(Axis) A non-positive integer indicating z-index of the axis. If zindex is 0, axes should be drawn behind all chart elements. To put them in front, use "zindex = 1".

Default value: 1 (in front of the marks) for actual axis and 0 (behind the marks) for grids.

Value

A modified Vega-Lite Spec