Config

Create object for Config

Arguments

area
(Config) Area-Specific Config

autosize
(Config) How the visualization size should be determined. If a string, should be one of "pad", "fit" or "none". Object values can additionally specify parameters for content sizing and automatic resizing.

Default value: pad

axis
(Config) Axis configuration, which determines default properties for all x and y axes. For a full list of axis configuration options, please see the corresponding section of the axis documentation.

axisBand
(Config) Specific axis config for axes with "band" scales.

axisBottom
(Config) Specific axis config for x-axis along the bottom edge of the chart.

axisLeft
(Config) Specific axis config for y-axis along the left edge of the chart.

axisRight
(Config) Specific axis config for y-axis along the right edge of the chart.

axisTop
(Config) Specific axis config for x-axis along the top edge of the chart.

axisX
(Config) X-axis specific config.

axisY
(Config) Y-axis specific config.

background
(Config) CSS color property to use as the background of the entire view.

Default value: "white"

bar
(Config) Bar-Specific Config

boxplot
(Config) Box Config

circle
(Config) Circle-Specific Config

concat
(Config) Default configuration for all concatenation view composition operators (concat, hconcat, and vconcat)

countTitle
(Config) Default axis and legend title for count fields.

Default value: Count of Records.

errorband
(Config) ErrorBand Config

errorbar
(Config) ErrorBar Config

facet
(Config) Default configuration for the facet view composition operator

fieldTitle
(Config) Defines how Vega-Lite generates title for fields. There are three possible styles:

"verbal" (Default) - displays function in a verbal style (e.g., "Sum of field", "Year-month of date", "field (binned)").

"function" - displays function using parentheses and capitalized texts (e.g., "SUM(field)", "YEARMONTH(date)", "BIN(field)").

"plain" - displays only the field name without functions (e.g., "field", "date", "field").

geoshape
(Config) Geoshape-Specific Config

header
(Config) Header configuration, which determines default properties for all headers.

For a full list of header configuration options, please see the corresponding section of in the header documentation.

headerColumn
(Config) Header configuration, which determines default properties for column headers.

For a full list of header configuration options, please see the corresponding section of in the header documentation.

headerFacet
(Config) Header configuration, which determines default properties for non-row/column facet headers.

For a full list of header configuration options, please see the corresponding section of in the header documentation.

headerRow
(Config) Header configuration, which determines default properties for row headers.

For a full list of header configuration options, please see the corresponding section of in the header documentation.

image
(Config) Image-specific Config

legend
(Config) Legend configuration, which determines default properties for all legends. For a full list of legend configuration options, please see the corresponding section of in the legend documentation.

line
(Config) Line-Specific Config

mark
(Config) Mark Config

numberFormat
(Config) D3 Number format for guide labels and text marks. For example "s" for SI units. Use D3's number format pattern.

padding
(Config) The default visualization padding, in pixels, from the edge of the visualization canvas to the data rectangle. If a number, specifies padding for all sides. If an object, the value should have the format {"left": 5, "top": 5, "right": 5, "bottom": 5} to specify padding for each side of the visualization.

Default value: 5

point
(Config) Point-Specific Config

projection
(Config) Projection configuration, which determines default properties for all projections. For a full list of projection configuration options, please see the corresponding section of the projection documentation.

range
(Config) An object hash that defines default range arrays or schemes for using with scales. For a full list of scale range configuration options, please see the corresponding section of the scale documentation.

rect
(Config) Rect-Specific Config

repeat
(Config) Default configuration for the repeat view composition operator

rule
(Config) Rule-Specific Config

scale
(Config) Scale configuration determines default properties for all scales. For a full list of scale configuration options, please see the corresponding section of the scale documentation.

selection
(Config) An object hash for defining default properties for each type of selections.

square
(Config) Square-Specific Config

style
(Config) An object hash that defines key-value mappings to determine default properties for marks with a given style. The keys represent styles names; the values have to be valid mark configuration objects.

text
(Config) Text-Specific Config

tick
(Config) Tick-Specific Config

timeFormat
(Config) Default time format for raw time values (without time units) in text marks, legend labels and header labels.

Default value: "%b %d, %Y" Note: Axes automatically determine format each label automatically so this config would not affect axes.

title
(Config) Title configuration, which determines default properties for all titles. For a full list of title configuration options, please see the corresponding section of the title documentation.

trail
(Config) Trail-Specific Config

view
(Config) Default properties for single view plots.

Value

A component of a Vega-Lite spec, corresponding to Config definition.