ProjectionConfig

Create object for ProjectionConfig

Arguments

center
(Projection) The projection’s center to the specified center, a two-element array of longitude and latitude in degrees.

Default value: [0, 0]

clipAngle
(Projection) The projection’s clipping circle radius to the specified angle in degrees. If null, switches to antimeridian cutting rather than small-circle clipping.

clipExtent
(Projection) The projection’s viewport clip extent to the specified bounds in pixels. The extent bounds are specified as an array [\[x0, y0\], \[x1, y1\]], where x0 is the left-side of the viewport, y0 is the top, x1 is the right and y1 is the bottom. If null, no viewport clipping is performed.

coefficient
(Projection)

distance
(Projection)

fraction
(Projection)

lobes
(Projection)

parallel
(Projection)

parallels
(Projection)

precision
(Projection) The threshold for the projection’s adaptive resampling to the specified value in pixels. This value corresponds to the Douglas–Peucker distance. If precision is not specified, returns the projection’s current resampling precision which defaults to √0.5 ≅ 0.70710….

radius
(Projection)

ratio
(Projection)

reflectX
(Projection)

reflectY
(Projection)

rotate
(Projection) The projection’s three-axis rotation to the specified angles, which must be a two- or three-element array of numbers [lambda, phi, gamma] specifying the rotation angles in degrees about each spherical axis. (These correspond to yaw, pitch and roll.)

Default value: [0, 0, 0]

scale
(Projection) The projection's scale (zoom) value, overriding automatic fitting.

spacing
(Projection)

tilt
(Projection)

translate
(Projection) The projection's translation (pan) value, overriding automatic fitting.

type
(Projection) The cartographic projection to use. This value is case-insensitive, for example "albers" and "Albers" indicate the same projection type. You can find all valid projection types in the documentation.

Default value: mercator

Value

A component of a Vega-Lite spec, corresponding to ProjectionConfig definition.