SingleSelectionConfig

Create object for SingleSelectionConfig

Arguments

bind
(SingleSelectionConfig) When set, a selection is populated by input elements (also known as dynamic query widgets) or by interacting with the corresponding legend. Direct manipulation interaction is disabled by default; to re-enable it, set the selection's on property.

Legend bindings are restricted to selections that only specify a single field or encoding.

Query widget binding takes the form of Vega's input element binding definition or can be a mapping between projected field/encodings and binding definitions.

See also: bind documentation.

clear
(SingleSelectionConfig) Clears the selection, emptying it of all values. Can be a Event Stream or false to disable.

Default value: dblclick.

See also: clear documentation.

empty
(SingleSelectionConfig) By default, all data values are considered to lie within an empty selection. When set to none, empty selections contain no data values.

encodings
(SingleSelectionConfig) An array of encoding channels. The corresponding data field values must match for a data tuple to fall within the selection.

See also: encodings documentation.

fields
(SingleSelectionConfig) An array of field names whose values must match for a data tuple to fall within the selection.

See also: fields documentation.

init
(SingleSelectionConfig) Initialize the selection with a mapping between projected channels or field names and initial values.

See also: init documentation.

nearest
(SingleSelectionConfig) When true, an invisible voronoi diagram is computed to accelerate discrete selection. The data value nearest the mouse cursor is added to the selection.

See also: nearest documentation.

on
(SingleSelectionConfig) A Vega event stream (object or selector) that triggers the selection. For interval selections, the event stream must specify a start and end.

resolve
(SingleSelectionConfig) With layered and multi-view displays, a strategy that determines how selections' data queries are resolved when applied in a filter transform, conditional encoding rule, or scale domain.

See also: resolve documentation.

Value

A component of a Vega-Lite spec, corresponding to SingleSelectionConfig definition.