Url

Create object for Url

Arguments

aggregate
(FieldDefWithCondition<StringFieldDef,string>) Aggregation function for the field (e.g., "mean", "sum", "median", "min", "max", "count").

Default value: undefined (None)

See also: aggregate documentation.

bin
(FieldDefWithCondition<StringFieldDef,string>) A flag for binning a quantitative field, an object defining binning parameters, or indicating that the data for x or y channel are binned before they are imported into Vega-Lite ("binned").

If true, default binning parameters will be applied.

If "binned", this indicates that the data for the x (or y) channel are already binned. You can map the bin-start field to x (or y) and the bin-end field to x2 (or y2). The scale and axis will be formatted similar to binning in Vega-Lite. To adjust the axis ticks based on the bin step, you can also set the axis's tickMinStep property.

Default value: false

See also: bin documentation.

condition
(ValueDefWithCondition<MarkPropFieldDef,(string|null)>) A field definition or one or more value definition(s) with a selection predicate.

(FieldDefWithCondition<StringFieldDef,string>) One or more value definition(s) with a selection or a test predicate.

Note: A field definition's condition property can only contain conditional value definitions since Vega-Lite only allows at most one encoded field per encoding channel.

field
(FieldDefWithCondition<StringFieldDef,string>) Required. A string defining the name of the field from which to pull a data value or an object defining iterated values from the repeat operator.

See also: field documentation.

Notes:

Dots (.) and brackets ([ and ]) can be used to access nested objects (e.g., "field": "foo.bar" and "field": "foo\['bar'\]"). If field names contain dots or brackets but are not nested, you can use \\ to escape dots and brackets (e.g., "a\\.b" and "a\\\[0\\\]"). See more details about escaping in the field documentation.

field is not required if aggregate is count.

format
(FieldDefWithCondition<StringFieldDef,string>) The text formatting pattern for labels of guides (axes, legends, headers) and text marks.

If the format type is "number" (e.g., for quantitative fields), this is D3's number format pattern.

If the format type is "time" (e.g., for temporal fields), this is D3's time format pattern.

See the format documentation for more examples.

Default value: Derived from numberFormat config for number format and from timeFormat config for time format.

formatType
(FieldDefWithCondition<StringFieldDef,string>) The format type for labels ("number" or "time").

Default value:

"time" for temporal fields and ordinal and nomimal fields with timeUnit.

"number" for quantitative fields as well as ordinal and nomimal fields without timeUnit.

labelExpr
(FieldDefWithCondition<StringFieldDef,string>) Vega expression for customizing labels text.

Note: The label text and value can be assessed via the label and value properties of the axis's backing datum object.

timeUnit
(FieldDefWithCondition<StringFieldDef,string>) Time unit (e.g., year, yearmonth, month, hours) for a temporal field. or a temporal field that gets casted as ordinal.

Default value: undefined (None)

See also: timeUnit documentation.

title
(FieldDefWithCondition<StringFieldDef,string>) A title for the field. If null, the title will be removed.

Default value: derived from the field's name and transformation function (aggregate, bin and timeUnit). If the field has an aggregate function, the function is displayed as part of the title (e.g., "Sum of Profit"). If the field is binned or has a time unit applied, the applied function is shown in parentheses (e.g., "Profit (binned)", "Transaction Date (year-month)"). Otherwise, the title is simply the field name.

Notes:

You can customize the default field title format by providing the fieldTitle property in the config or fieldTitle function via the compile function's options.

If both field definition's title and axis, header, or legend title are defined, axis/header/legend title will be used.

type
(FieldDefWithCondition<StringFieldDef,string>) The encoded field's type of measurement ("quantitative", "temporal", "ordinal", or "nominal"). It can also be a "geojson" type for encoding 'geoshape'.

Note:

Data values for a temporal field can be either a date-time string (e.g., "2015-03-07 12:32:17", "17:01", "2015-03-16". "2015") or a timestamp number (e.g., 1552199579097).

Data type describes the semantics of the data rather than the primitive data types (number, string, etc.). The same primitive data type can have different types of measurement. For example, numeric data can represent quantitative, ordinal, or nominal data.

When using with bin, the type property can be either "quantitative" (for using a linear bin scale) or "ordinal" (for using an ordinal bin scale).

When using with timeUnit, the type property can be either "temporal" (for using a temporal scale) or "ordinal" (for using an ordinal scale).

When using with aggregate, the type property refers to the post-aggregation data type. For example, we can calculate count distinct of a categorical field "cat" using {"aggregate": "distinct", "field": "cat", "type": "quantitative"}. The "type" of the aggregate output is "quantitative".

Secondary channels (e.g., x2, y2, xError, yError) do not have type as they have exactly the same type as their primary channels (e.g., x, y).

See also: type documentation.

value
(ValueDefWithCondition<MarkPropFieldDef,(string|null)>) A constant value in visual domain (e.g., "red" / "#0099ff" / gradient definition for color, values between 0 to 1 for opacity).

Value

A component of a Vega-Lite spec, corresponding to Url definition.