ViewConfig

Create object for ViewConfig

Arguments

clip
(ViewConfig) Whether the view should be clipped.

continuousHeight
(ViewConfig) The default height when the plot has a continuous y-field.

Default value: 200

continuousWidth
(ViewConfig) The default width when the plot has a continuous x-field.

Default value: 200

cornerRadius
(ViewConfig) The radius in pixels of rounded rectangle corners.

Default value: 0

discreteHeight
(ViewConfig) The default height when the plot has either a discrete y-field or no y-field.

Default value: a step size based on config.view.step.

discreteWidth
(ViewConfig) The default width when the plot has either a discrete x-field or no x-field.

Default value: a step size based on config.view.step.

fill
(ViewConfig) The fill color.

Default value: undefined

fillOpacity
(ViewConfig) The fill opacity (value between [0,1]).

Default value: 1

height
(ViewConfig) Default height

opacity
(ViewConfig) The overall opacity (value between [0,1]).

Default value: 0.7 for non-aggregate plots with point, tick, circle, or square marks or layered bar charts and 1 otherwise.

step
(ViewConfig) Default step size for x-/y- discrete fields.

stroke
(ViewConfig) The stroke color.

Default value: "#ddd"

strokeCap
(ViewConfig) The stroke cap for line ending style. One of "butt", "round", or "square".

Default value: "square"

strokeDash
(ViewConfig) An array of alternating stroke, space lengths for creating dashed or dotted lines.

strokeDashOffset
(ViewConfig) The offset (in pixels) into which to begin drawing with the stroke dash array.

strokeJoin
(ViewConfig) The stroke line join method. One of "miter", "round" or "bevel".

Default value: "miter"

strokeMiterLimit
(ViewConfig) The miter limit at which to bevel a line join.

strokeOpacity
(ViewConfig) The stroke opacity (value between [0,1]).

Default value: 1

strokeWidth
(ViewConfig) The stroke width, in pixels.

width
(ViewConfig) Default width

Value

A component of a Vega-Lite spec, corresponding to ViewConfig definition.