vl_add_headerColumn_config

Add headerColumn config (HeaderConfig) to a vega-lite spec.

vl_add_headerColumn_config(spec, format = NULL, formatType = NULL,
  labelAlign = NULL, labelAnchor = NULL, labelAngle = NULL,
  labelColor = NULL, labelFont = NULL, labelFontSize = NULL,
  labelLimit = NULL, labelOrient = NULL, labelPadding = NULL,
  shortTimeLabels = NULL, title = NULL, titleAlign = NULL,
  titleAnchor = NULL, titleAngle = NULL, titleBaseline = NULL,
  titleColor = NULL, titleFont = NULL, titleFontSize = NULL,
  titleFontWeight = NULL, titleLimit = NULL, titleOrient = NULL,
  titlePadding = NULL)

Arguments

spec
An input vega-lite spec

format
(HeaderConfig) The text formatting pattern for labels of guides (axes, legends, headers) and text marks.

If the format type is "number" (e.g., for quantitative fields), this is D3's number format pattern.

If the format type is "time" (e.g., for temporal fields), this is D3's time format pattern.

See the format documentation for more examples.

Default value: Derived from numberFormat config for number format and from timeFormat config for time format.

formatType
(HeaderConfig) The format type for labels ("number" or "time").

Default value:

"time" for temporal fields and ordinal and nomimal fields with timeUnit.

"number" for quantitative fields as well as ordinal and nomimal fields without timeUnit.

labelAlign
(HeaderConfig) Horizontal text alignment of header labels.

labelAnchor
(HeaderConfig) The anchor position for placing the labels. One of "start", "middle", or "end". For example, with a label orientation of top these anchor positions map to a left-, center-, or right-aligned label.

labelAngle
(HeaderConfig) The rotation angle of the header labels.

Default value: 0 for column header, -90 for row header.

labelColor
(HeaderConfig) The color of the header label, can be in hex color code or regular color name.

labelFont
(HeaderConfig) The font of the header label.

labelFontSize
(HeaderConfig) The font size of the header label, in pixels.

labelLimit
(HeaderConfig) The maximum length of the header label in pixels. The text value will be automatically truncated if the rendered size exceeds the limit.

Default value: 0, indicating no limit

labelOrient
(HeaderConfig) The orientation of the header label. One of "top", "bottom", "left" or "right".

labelPadding
(HeaderConfig) The padding, in pixel, between facet header's label and the plot.

Default value: 10

shortTimeLabels
(HeaderConfig) Whether month names and weekday names should be abbreviated.

Default value: false

title
(HeaderConfig) Set to null to disable title for the axis, legend, or header.

titleAlign
(HeaderConfig) Horizontal text alignment (to the anchor) of header titles.

titleAnchor
(HeaderConfig) The anchor position for placing the title. One of "start", "middle", or "end". For example, with an orientation of top these anchor positions map to a left-, center-, or right-aligned title.

titleAngle
(HeaderConfig) The rotation angle of the header title.

Default value: 0.

titleBaseline
(HeaderConfig) Vertical text baseline for the header title. One of "top", "bottom", "middle".

Default value: "middle"

titleColor
(HeaderConfig) Color of the header title, can be in hex color code or regular color name.

titleFont
(HeaderConfig) Font of the header title. (e.g., "Helvetica Neue").

titleFontSize
(HeaderConfig) Font size of the header title.

titleFontWeight
(HeaderConfig) Font weight of the header title. This can be either a string (e.g "bold", "normal") or a number (100, 200, 300, ..., 900 where "normal" = 400 and "bold" = 700).

titleLimit
(HeaderConfig) The maximum length of the header title in pixels. The text value will be automatically truncated if the rendered size exceeds the limit.

Default value: 0, indicating no limit

titleOrient
(HeaderConfig) The orientation of the header title. One of "top", "bottom", "left" or "right".

titlePadding
(HeaderConfig) The padding, in pixel, between facet header's title and the label.

Default value: 10

Value

A modified Vega-Lite Spec