vl_add_interval_selection

Add IntervalSelection to a vega-lite spec.

vl_add_interval_selection(spec, selection_name, .object = NULL,
  bind = NULL, clear = NULL, empty = NULL, encodings = NULL,
  fields = NULL, init = NULL, mark = NULL, on = NULL,
  resolve = NULL, translate = NULL, type = NULL, zoom = NULL)

Arguments

spec
An input vega-lite spec

selection_name
Name for selection

.object
Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: IntervalSelection

bind
(IntervalSelection) Establishes a two-way binding between the interval selection and the scales used within the same view. This allows a user to interactively pan and zoom the view.

See also: bind documentation.

clear
(IntervalSelection) Clears the selection, emptying it of all values. Can be a Event Stream or false to disable.

Default value: dblclick.

See also: clear documentation.

empty
(IntervalSelection) By default, all data values are considered to lie within an empty selection. When set to none, empty selections contain no data values.

encodings
(IntervalSelection) An array of encoding channels. The corresponding data field values must match for a data tuple to fall within the selection.

See also: encodings documentation.

fields
(IntervalSelection) An array of field names whose values must match for a data tuple to fall within the selection.

See also: fields documentation.

init
(IntervalSelection) Initialize the selection with a mapping between projected channels or field names and arrays of initial values.

See also: init documentation.

mark
(IntervalSelection) An interval selection also adds a rectangle mark to depict the extents of the interval. The mark property can be used to customize the appearance of the mark.

See also: mark documentation.

on
(IntervalSelection) A Vega event stream (object or selector) that triggers the selection. For interval selections, the event stream must specify a start and end.

resolve
(IntervalSelection) With layered and multi-view displays, a strategy that determines how selections' data queries are resolved when applied in a filter transform, conditional encoding rule, or scale domain.

See also: resolve documentation.

translate
(IntervalSelection) When truthy, allows a user to interactively move an interval selection back-and-forth. Can be true, false (to disable panning), or a Vega event stream definition which must include a start and end event to trigger continuous panning.

Default value: true, which corresponds to [mousedown, window:mouseup] > window:mousemove! which corresponds to clicks and dragging within an interval selection to reposition it.

See also: translate documentation.

type
(IntervalSelection) Determines the default event processing and data query for the selection. Vega-Lite currently supports three selection types:

"single" -- to select a single discrete data value on click.

"multi" -- to select multiple discrete data value; the first value is selected on click and additional values toggled on shift-click.

"interval" -- to select a continuous range of data values on drag.

zoom
(IntervalSelection) When truthy, allows a user to interactively resize an interval selection. Can be true, false (to disable zooming), or a Vega event stream definition. Currently, only wheel events are supported.

Default value: true, which corresponds to wheel!.

See also: zoom documentation.

Value

A modified Vega-Lite Spec