vl_add_multi_selection

Add MultiSelection to a vega-lite spec.

vl_add_multi_selection(
  spec,
  selection_name,
  .object = NULL,
  bind = NULL,
  clear = NULL,
  empty = NULL,
  encodings = NULL,
  fields = NULL,
  init = NULL,
  nearest = NULL,
  on = NULL,
  resolve = NULL,
  toggle = NULL,
  type = NULL
)

Arguments

spec
An input vega-lite spec

selection_name
Name for selection

.object
Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: MultiSelection

bind
(MultiSelection) When set, a selection is populated by interacting with the corresponding legend. Direct manipulation interaction is disabled by default; to re-enable it, set the selection's on property.

Legend bindings are restricted to selections that only specify a single field or encoding.

clear
(MultiSelection) Clears the selection, emptying it of all values. Can be a Event Stream or false to disable.

Default value: dblclick.

See also: clear documentation.

empty
(MultiSelection) By default, all data values are considered to lie within an empty selection. When set to none, empty selections contain no data values.

encodings
(MultiSelection) An array of encoding channels. The corresponding data field values must match for a data tuple to fall within the selection.

See also: encodings documentation.

fields
(MultiSelection) An array of field names whose values must match for a data tuple to fall within the selection.

See also: fields documentation.

init
(MultiSelection) Initialize the selection with a mapping between projected channels or field names and an initial value (or array of values).

See also: init documentation.

nearest
(MultiSelection) When true, an invisible voronoi diagram is computed to accelerate discrete selection. The data value nearest the mouse cursor is added to the selection.

See also: nearest documentation.

on
(MultiSelection) A Vega event stream (object or selector) that triggers the selection. For interval selections, the event stream must specify a start and end.

resolve
(MultiSelection) With layered and multi-view displays, a strategy that determines how selections' data queries are resolved when applied in a filter transform, conditional encoding rule, or scale domain.

See also: resolve documentation.

toggle
(MultiSelection) Controls whether data values should be toggled or only ever inserted into multi selections. Can be true, false (for insertion only), or a Vega expression.

Default value: true, which corresponds to event.shiftKey (i.e., data values are toggled when a user interacts with the shift-key pressed).

See also: toggle documentation.

type
(MultiSelection) Determines the default event processing and data query for the selection. Vega-Lite currently supports three selection types:

"single" -- to select a single discrete data value on click.

"multi" -- to select multiple discrete data value; the first value is selected on click and additional values toggled on shift-click.

"interval" -- to select a continuous range of data values on drag.

Value

A modified Vega-Lite Spec