vl_bin_size

Add bin to encoding for size in a vega-lite spec.

vl_bin_size(spec, anchor = NULL, base = NULL, binned = NULL,
  divide = NULL, extent = NULL, maxbins = NULL, minstep = NULL,
  nice = NULL, step = NULL, steps = NULL)

Arguments

spec
A vega-lite spec.

anchor
A value in the binned domain at which to anchor the bins, shifting the bin boundaries if necessary to ensure that a boundary aligns with the anchor value.

Default Value: the minimum bin extent value (type = number)

base
The number base to use for automatic bin determination (default is base 10).

Default value: 10 (type = number)

binned
When set to true, Vega-Lite treats the input data as already binned. (type = boolean)

divide
Scale factors indicating allowable subdivisions. The default value is 5, 2, which indicates that for base 10 numbers (the default base), the method may consider dividing bin sizes by 5 and/or 2. For example, for an initial step size of 10, the method can check if bin sizes of 2 (= 10/5), 5 (= 10/2), or 1 (= 10/(5*2)) might also satisfy the given constraints.

Default value: [5, 2] (type = array)

extent
A two-element ([min, max]) array indicating the range of desired bin values. (type = array)

maxbins
Maximum number of bins.

Default value: 6 for row, column and shape channels; 10 for other channels (type = number)

minstep
A minimum allowable step size (particularly useful for integer values). (type = number)

nice
If true (the default), attempts to make the bin boundaries use human-friendly boundaries, such as multiples of ten. (type = boolean)

step
An exact step size to use between bins.

Note: If provided, options such as maxbins will be ignored. (type = number)

steps
An array of allowable step sizes to choose from. (type = array)

Value

A modified spec

See also