vl_bind_checkbox_input

Add checkbox binding to a vega-lite spec.

vl_bind_checkbox_input(
  spec,
  parameter_name,
  projection_name = NULL,
  .object = NULL,
  debounce = NULL,
  element = NULL,
  name = NULL
)

Arguments

spec
An input vega-lite spec

parameter_name
Name of selection to add binding to

projection_name
Name of projection (field or encoding) within selection

.object
Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: BindCheckbox

debounce
If defined, delays event handling until the specified milliseconds have elapsed since the last event was fired.

element
An optional CSS selector string indicating the parent element to which the input element should be added. By default, all input elements are added within the parent container of the Vega view.

name
By default, the signal name is used to label input elements. This name property can be used instead to specify a custom label for the bound signal.

Value

A modified Vega-Lite Spec