vl_bind_range_input

Add range binding to a vega-lite spec.

Usage

vl_bind_range_input(
  spec,
  parameter_name,
  projection_name = NULL,
  debounce = NULL,
  element = NULL,
  max = NULL,
  min = NULL,
  name = NULL,
  step = NULL,
  .object = NULL
)

Arguments

spec

An input vega-lite spec

parameter_name

Name of parameter to add binding to

projection_name

Name of projection (field or encoding) within selection

debounce

If defined, delays event handling until the specified milliseconds have elapsed since the last event was fired.

element

An optional CSS selector string indicating the parent element to which the input element should be added. By default, all input elements are added within the parent container of the Vega view.

max

Sets the maximum slider value. Defaults to the larger of the signal value and 100.

min

Sets the minimum slider value. Defaults to the smaller of the signal value and 0.

name

By default, the signal name is used to label input elements. This name property can be used instead to specify a custom label for the bound signal.

step

Sets the minimum slider increment. If undefined, the step size will be automatically determined based on the min and max values.

.object

Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than spec. Objects can be of type: BindRange

Value

A modified Vega-Lite Spec