vl_condition_size

Add condition to encoding for size in a vega-lite spec.

vl_condition_size(spec, test = NULL, value = NULL, selection = NULL,
  aggregate = NULL, bin = NULL, field = NULL, legend = NULL,
  scale = NULL, sort = NULL, timeUnit = NULL, title = NULL,
  type = NULL)

Arguments

spec
A vega-lite spec.

test
(ConditionalPredicate<ValueDef>, ConditionalPredicate<MarkPropFieldDef>) Predicate for triggering the condition

value
(ConditionalPredicate<ValueDef>, ConditionalSelection<ValueDef>) A constant value in visual domain (e.g., "red" / "#0099ff" for color, values between 0 to 1 for opacity).

selection
(ConditionalSelection<ValueDef>, ConditionalSelection<MarkPropFieldDef>) A selection name, or a series of composed selections.

aggregate
(ConditionalPredicate<MarkPropFieldDef>, ConditionalSelection<MarkPropFieldDef>) Aggregation function for the field (e.g., mean, sum, median, min, max, count).

Default value: undefined (None)

bin
(ConditionalPredicate<MarkPropFieldDef>, ConditionalSelection<MarkPropFieldDef>) A flag for binning a quantitative field, an object defining binning parameters, or indicating that the data for x or y channel are binned before they are imported into Vega-Lite ("binned").

If true, default binning parameters will be applied.

If "binned", this indicates that the data for the x (or y) channel are already binned. You can map the bin-start field to x (or y) and the bin-end field to x2 (or y2). The scale and axis will be formatted similar to binning in Vega-lite. To adjust the axis ticks based on the bin step, you can also set the axis's tickMinStep property.

Default value: false

field
(ConditionalPredicate<MarkPropFieldDef>, ConditionalSelection<MarkPropFieldDef>) Required. A string defining the name of the field from which to pull a data value or an object defining iterated values from the repeat operator.

Note: Dots (.) and brackets ([ and ]) can be used to access nested objects (e.g., "field": "foo.bar" and "field": "foo['bar']"). If field names contain dots or brackets but are not nested, you can use \ to escape dots and brackets (e.g., "a\.b" and "a\[0\]"). See more details about escaping in the field documentation.

Note: field is not required if aggregate is count.

legend
(ConditionalPredicate<MarkPropFieldDef>, ConditionalSelection<MarkPropFieldDef>) An object defining properties of the legend. If null, the legend for the encoding channel will be removed.

Default value: If undefined, default legend properties are applied.

scale
(ConditionalPredicate<MarkPropFieldDef>, ConditionalSelection<MarkPropFieldDef>) An object defining properties of the channel's scale, which is the function that transforms values in the data domain (numbers, dates, strings, etc) to visual values (pixels, colors, sizes) of the encoding channels.

If null, the scale will be disabled and the data value will be directly encoded.

Default value: If undefined, default scale properties are applied.

sort
(ConditionalPredicate<MarkPropFieldDef>, ConditionalSelection<MarkPropFieldDef>) Sort order for the encoded field.

For continuous fields (quantitative or temporal), sort can be either "ascending" or "descending".

For discrete fields, sort can be one of the following:

"ascending" or "descending" -- for sorting by the values' natural order in Javascript.

A sort-by-encoding definition for sorting by another encoding channel. (This type of sort definition is not available for row and column channels.)

A sort field definition for sorting by another field.

An array specifying the field values in preferred order. In this case, the sort order will obey the values in the array, followed by any unspecified values in their original order. For discrete time field, values in the sort array can be date-time definition objects. In addition, for time units "month" and "day", the values can be the month or day names (case insensitive) or their 3-letter initials (e.g., "Mon", "Tue").

null indicating no sort.

Default value: "ascending"

Note: null is not supported for row and column.

timeUnit
(ConditionalPredicate<MarkPropFieldDef>, ConditionalSelection<MarkPropFieldDef>) Time unit (e.g., year, yearmonth, month, hours) for a temporal field. or a temporal field that gets casted as ordinal.

Default value: undefined (None)

title
(ConditionalPredicate<MarkPropFieldDef>, ConditionalSelection<MarkPropFieldDef>) A title for the field. If null, the title will be removed.

Default value: derived from the field's name and transformation function (aggregate, bin and timeUnit). If the field has an aggregate function, the function is displayed as part of the title (e.g., "Sum of Profit"). If the field is binned or has a time unit applied, the applied function is shown in parentheses (e.g., "Profit (binned)", "Transaction Date (year-month)"). Otherwise, the title is simply the field name.

Notes:

You can customize the default field title format by providing the fieldTitle property in the config or fieldTitle function via the compile function's options.

If both field definition's title and axis, header, or legend title are defined, axis/header/legend title will be used.

type
(ConditionalPredicate<MarkPropFieldDef>, ConditionalSelection<MarkPropFieldDef>) The encoded field's type of measurement ("quantitative", "temporal", "ordinal", or "nominal"). It can also be a "geojson" type for encoding 'geoshape'.

Note:

Data values for a temporal field can be either a date-time string (e.g., "2015-03-07 12:32:17", "17:01", "2015-03-16". "2015") or a timestamp number (e.g., 1552199579097).

Data type describes the semantics of the data rather than the primitive data types (number, string, etc.). The same primitive data type can have different types of measurement. For example, numeric data can represent quantitative, ordinal, or nominal data.

When using with bin, the type property can be either "quantitative" (for using a linear bin scale) or "ordinal" (for using an ordinal bin scale).

When using with timeUnit, the type property can be either "temporal" (for using a temporal scale) or "ordinal" (for using an ordinal scale).

When using with aggregate, the type property refers to the post-aggregation data type. For example, we can calculate count distinct of a categorical field "cat" using {"aggregate": "distinct", "field": "cat", "type": "quantitative"}. The "type" of the aggregate output is "quantitative".

Secondary channels (e.g., x2, y2, xError, yError) do not have type as they have exactly the same type as their primary channels (e.g., x, y).

Value

A modified spec

See also