vl_config

Add top-level config to a vega-lite spec.

vl_config(
  spec,
  .object = NULL,
  arc = NULL,
  area = NULL,
  aria = NULL,
  autosize = NULL,
  axis = NULL,
  axisBand = NULL,
  axisBottom = NULL,
  axisDiscrete = NULL,
  axisLeft = NULL,
  axisPoint = NULL,
  axisQuantitative = NULL,
  axisRight = NULL,
  axisTemporal = NULL,
  axisTop = NULL,
  axisX = NULL,
  axisXBand = NULL,
  axisXDiscrete = NULL,
  axisXPoint = NULL,
  axisXQuantitative = NULL,
  axisXTemporal = NULL,
  axisY = NULL,
  axisYBand = NULL,
  axisYDiscrete = NULL,
  axisYPoint = NULL,
  axisYQuantitative = NULL,
  axisYTemporal = NULL,
  background = NULL,
  bar = NULL,
  boxplot = NULL,
  circle = NULL,
  concat = NULL,
  countTitle = NULL,
  customFormatTypes = NULL,
  errorband = NULL,
  errorbar = NULL,
  facet = NULL,
  fieldTitle = NULL,
  font = NULL,
  geoshape = NULL,
  header = NULL,
  headerColumn = NULL,
  headerFacet = NULL,
  headerRow = NULL,
  image = NULL,
  legend = NULL,
  line = NULL,
  lineBreak = NULL,
  locale = NULL,
  mark = NULL,
  numberFormat = NULL,
  padding = NULL,
  params = NULL,
  point = NULL,
  projection = NULL,
  range = NULL,
  rect = NULL,
  rule = NULL,
  scale = NULL,
  selection = NULL,
  square = NULL,
  style = NULL,
  text = NULL,
  tick = NULL,
  timeFormat = NULL,
  title = NULL,
  trail = NULL,
  view = NULL
)

Arguments

spec
An input vega-lite spec

.object
Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: Config

arc
Arc-specific Config

area
Area-Specific Config

aria
A boolean flag indicating if ARIA default attributes should be included for marks and guides (SVG output only). If false, the "aria-hidden" attribute will be set for all guides, removing them from the ARIA accessibility tree and Vega-Lite will not generate default descriptions for marks.

Default value: true.

autosize
How the visualization size should be determined. If a string, should be one of "pad", "fit" or "none". Object values can additionally specify parameters for content sizing and automatic resizing.

Default value: pad

axis
Axis configuration, which determines default properties for all x and y axes. For a full list of axis configuration options, please see the corresponding section of the axis documentation.

axisBand
Config for axes with "band" scales.

axisBottom
Config for x-axis along the bottom edge of the chart.

axisDiscrete
Config for axes with "point" or "band" scales.

axisLeft
Config for y-axis along the left edge of the chart.

axisPoint
Config for axes with "point" scales.

axisQuantitative
Config for quantitative axes.

axisRight
Config for y-axis along the right edge of the chart.

axisTemporal
Config for temporal axes.

axisTop
Config for x-axis along the top edge of the chart.

axisX
X-axis specific config.

axisXBand
Config for x-axes with "band" scales.

axisXDiscrete
Config for x-axes with "point" or "band" scales.

axisXPoint
Config for x-axes with "point" scales.

axisXQuantitative
Config for x-quantitative axes.

axisXTemporal
Config for x-temporal axes.

axisY
Y-axis specific config.

axisYBand
Config for y-axes with "band" scales.

axisYDiscrete
Config for y-axes with "point" or "band" scales.

axisYPoint
Config for y-axes with "point" scales.

axisYQuantitative
Config for y-quantitative axes.

axisYTemporal
Config for y-temporal axes.

background
CSS color property to use as the background of the entire view.

Default value: "white"

bar
Bar-Specific Config

boxplot
Box Config

circle
Circle-Specific Config

concat
Default configuration for all concatenation and repeat view composition operators (concat, hconcat, vconcat, and repeat)

countTitle
Default axis and legend title for count fields.

Default value: Count of Records.

customFormatTypes
Allow the formatType property for text marks and guides to accept a custom formatter function registered as a Vega expression.

errorband
ErrorBand Config

errorbar
ErrorBar Config

facet
Default configuration for the facet view composition operator

fieldTitle
Defines how Vega-Lite generates title for fields. There are three possible styles:

"verbal" (Default) - displays function in a verbal style (e.g., "Sum of field", "Year-month of date", "field (binned)").

"function" - displays function using parentheses and capitalized texts (e.g., "SUM(field)", "YEARMONTH(date)", "BIN(field)").

"plain" - displays only the field name without functions (e.g., "field", "date", "field").

font
Default font for all text marks, titles, and labels.

geoshape
Geoshape-Specific Config

header
Header configuration, which determines default properties for all headers.

For a full list of header configuration options, please see the corresponding section of in the header documentation.

headerColumn
Header configuration, which determines default properties for column headers.

For a full list of header configuration options, please see the corresponding section of in the header documentation.

headerFacet
Header configuration, which determines default properties for non-row/column facet headers.

For a full list of header configuration options, please see the corresponding section of in the header documentation.

headerRow
Header configuration, which determines default properties for row headers.

For a full list of header configuration options, please see the corresponding section of in the header documentation.

image
Image-specific Config

legend
Legend configuration, which determines default properties for all legends. For a full list of legend configuration options, please see the corresponding section of in the legend documentation.

line
Line-Specific Config

lineBreak
A delimiter, such as a newline character, upon which to break text strings into multiple lines. This property provides a global default for text marks, which is overridden by mark or style config settings, and by the lineBreak mark encoding channel. If signal-valued, either string or regular expression (regexp) values are valid.

locale
Locale definitions for string parsing and formatting of number and date values. The locale object should contain number and/or time properties with locale definitions. Locale definitions provided in the config block may be overridden by the View constructor locale option.

mark
Mark Config

numberFormat
D3 Number format for guide labels and text marks. For example "s" for SI units. Use D3's number format pattern.

padding
The default visualization padding, in pixels, from the edge of the visualization canvas to the data rectangle. If a number, specifies padding for all sides. If an object, the value should have the format {"left": 5, "top": 5, "right": 5, "bottom": 5} to specify padding for each side of the visualization.

Default value: 5

params
Dynamic variables or selections that parameterize a visualization.

point
Point-Specific Config

projection
Projection configuration, which determines default properties for all projections. For a full list of projection configuration options, please see the corresponding section of the projection documentation.

range
An object hash that defines default range arrays or schemes for using with scales. For a full list of scale range configuration options, please see the corresponding section of the scale documentation.

rect
Rect-Specific Config

rule
Rule-Specific Config

scale
Scale configuration determines default properties for all scales. For a full list of scale configuration options, please see the corresponding section of the scale documentation.

selection
An object hash for defining default properties for each type of selections.

square
Square-Specific Config

style
An object hash that defines key-value mappings to determine default properties for marks with a given style. The keys represent styles names; the values have to be valid mark configuration objects.

text
Text-Specific Config

tick
Tick-Specific Config

timeFormat
Default time format for raw time values (without time units) in text marks, legend labels and header labels.

Default value: "%b %d, %Y" Note: Axes automatically determine the format for each label automatically so this config does not affect axes.

title
Title configuration, which determines default properties for all titles. For a full list of title configuration options, please see the corresponding section of the title documentation.

trail
Trail-Specific Config

view
Default properties for single view plots.

Value

A modified Vega-Lite Spec