vl_config_axis

Add axis config (AxisConfig) to a vega-lite spec.

vl_config_axis(
  spec,
  .object = NULL,
  bandPosition = NULL,
  domain = NULL,
  domainColor = NULL,
  domainDash = NULL,
  domainDashOffset = NULL,
  domainOpacity = NULL,
  domainWidth = NULL,
  grid = NULL,
  gridColor = NULL,
  gridDash = NULL,
  gridDashOffset = NULL,
  gridOpacity = NULL,
  gridWidth = NULL,
  labelAlign = NULL,
  labelAngle = NULL,
  labelBaseline = NULL,
  labelBound = NULL,
  labelColor = NULL,
  labelFlush = NULL,
  labelFlushOffset = NULL,
  labelFont = NULL,
  labelFontSize = NULL,
  labelFontStyle = NULL,
  labelFontWeight = NULL,
  labelLimit = NULL,
  labelOpacity = NULL,
  labelOverlap = NULL,
  labelPadding = NULL,
  labelSeparation = NULL,
  labels = NULL,
  maxExtent = NULL,
  minExtent = NULL,
  orient = NULL,
  tickBand = NULL,
  tickColor = NULL,
  tickDash = NULL,
  tickDashOffset = NULL,
  tickExtra = NULL,
  tickOffset = NULL,
  tickOpacity = NULL,
  tickRound = NULL,
  tickSize = NULL,
  tickWidth = NULL,
  ticks = NULL,
  title = NULL,
  titleAlign = NULL,
  titleAnchor = NULL,
  titleAngle = NULL,
  titleBaseline = NULL,
  titleColor = NULL,
  titleFont = NULL,
  titleFontSize = NULL,
  titleFontStyle = NULL,
  titleFontWeight = NULL,
  titleLimit = NULL,
  titleLineHeight = NULL,
  titleOpacity = NULL,
  titlePadding = NULL,
  titleX = NULL,
  titleY = NULL,
  translate = NULL
)

Arguments

spec
An input vega-lite spec

.object
Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: AxisConfig

bandPosition
(AxisConfig) An interpolation fraction indicating where, for band scales, axis ticks should be positioned. A value of 0 places ticks at the left edge of their bands. A value of 0.5 places ticks in the middle of their bands.

Default value: 0.5

domain
(AxisConfig) A boolean flag indicating if the domain (the axis baseline) should be included as part of the axis.

Default value: true

domainColor
(AxisConfig) Color of axis domain line.

Default value: "gray".

domainDash
(AxisConfig) An array of alternating [stroke, space] lengths for dashed domain lines.

domainDashOffset
(AxisConfig) The pixel offset at which to start drawing with the domain dash array.

domainOpacity
(AxisConfig) Opacity of the axis domain line.

domainWidth
(AxisConfig) Stroke width of axis domain line

Default value: 1

grid
(AxisConfig) A boolean flag indicating if grid lines should be included as part of the axis

Default value: true for continuous scales that are not binned; otherwise, false.

gridColor
(AxisConfig)

gridDash
(AxisConfig)

gridDashOffset
(AxisConfig)

gridOpacity
(AxisConfig)

gridWidth
(AxisConfig)

labelAlign
(AxisConfig)

labelAngle
(AxisConfig) The rotation angle of the axis labels.

Default value: -90 for nominal and ordinal fields; 0 otherwise.

labelBaseline
(AxisConfig)

labelBound
(AxisConfig) Indicates if labels should be hidden if they exceed the axis range. If false (the default) no bounds overlap analysis is performed. If true, labels will be hidden if they exceed the axis range by more than 1 pixel. If this property is a number, it specifies the pixel tolerance: the maximum amount by which a label bounding box may exceed the axis range.

Default value: false.

labelColor
(AxisConfig)

labelFlush
(AxisConfig) Indicates if the first and last axis labels should be aligned flush with the scale range. Flush alignment for a horizontal axis will left-align the first label and right-align the last label. For vertical axes, bottom and top text baselines are applied instead. If this property is a number, it also indicates the number of pixels by which to offset the first and last labels; for example, a value of 2 will flush-align the first and last labels and also push them 2 pixels outward from the center of the axis. The additional adjustment can sometimes help the labels better visually group with corresponding axis ticks.

Default value: true for axis of a continuous x-scale. Otherwise, false.

labelFlushOffset
(AxisConfig) Indicates the number of pixels by which to offset flush-adjusted labels. For example, a value of 2 will push flush-adjusted labels 2 pixels outward from the center of the axis. Offsets can help the labels better visually group with corresponding axis ticks.

Default value: 0.

labelFont
(AxisConfig)

labelFontSize
(AxisConfig)

labelFontStyle
(AxisConfig)

labelFontWeight
(AxisConfig)

labelLimit
(AxisConfig) Maximum allowed pixel width of axis tick labels.

Default value: 180

labelOpacity
(AxisConfig)

labelOverlap
(AxisConfig) The strategy to use for resolving overlap of axis labels. If false (the default), no overlap reduction is attempted. If set to true or "parity", a strategy of removing every other label is used (this works well for standard linear axes). If set to "greedy", a linear scan of the labels is performed, removing any labels that overlaps with the last visible label (this often works better for log-scaled axes).

Default value: true for non-nominal fields with non-log scales; "greedy" for log scales; otherwise false.

labelPadding
(AxisConfig) The padding, in pixels, between axis and text labels.

Default value: 2

labelSeparation
(AxisConfig) The minimum separation that must be between label bounding boxes for them to be considered non-overlapping (default 0). This property is ignored if labelOverlap resolution is not enabled.

labels
(AxisConfig) A boolean flag indicating if labels should be included as part of the axis.

Default value: true.

maxExtent
(AxisConfig) The maximum extent in pixels that axis ticks and labels should use. This determines a maximum offset value for axis titles.

Default value: undefined.

minExtent
(AxisConfig) The minimum extent in pixels that axis ticks and labels should use. This determines a minimum offset value for axis titles.

Default value: 30 for y-axis; undefined for x-axis.

orient
(AxisConfig) The orientation of the axis. One of "top", "bottom", "left" or "right". The orientation can be used to further specialize the axis type (e.g., a y-axis oriented towards the right edge of the chart).

Default value: "bottom" for x-axes and "left" for y-axes.

tickBand
(AxisConfig) For band scales, indicates if ticks and grid lines should be placed at the center of a band (default) or at the band extents to indicate intervals.

tickColor
(AxisConfig)

tickDash
(AxisConfig)

tickDashOffset
(AxisConfig)

tickExtra
(AxisConfig) Boolean flag indicating if an extra axis tick should be added for the initial position of the axis. This flag is useful for styling axes for band scales such that ticks are placed on band boundaries rather in the middle of a band. Use in conjunction with "bandPosition": 1 and an axis "padding" value of 0.

tickOffset
(AxisConfig) Position offset in pixels to apply to ticks, labels, and gridlines.

tickOpacity
(AxisConfig)

tickRound
(AxisConfig) Boolean flag indicating if pixel position values should be rounded to the nearest integer.

Default value: true

tickSize
(AxisConfig) The size in pixels of axis ticks.

Default value: 5

tickWidth
(AxisConfig)

ticks
(AxisConfig) Boolean value that determines whether the axis should include ticks.

Default value: true

title
(AxisConfig) Set to null to disable title for the axis, legend, or header.

titleAlign
(AxisConfig) Horizontal text alignment of axis titles.

titleAnchor
(AxisConfig) Text anchor position for placing axis titles.

titleAngle
(AxisConfig) Angle in degrees of axis titles.

titleBaseline
(AxisConfig) Vertical text baseline for axis titles.

titleColor
(AxisConfig) Color of the title, can be in hex color code or regular color name.

titleFont
(AxisConfig) Font of the title. (e.g., "Helvetica Neue").

titleFontSize
(AxisConfig) Font size of the title.

titleFontStyle
(AxisConfig) Font style of the title.

titleFontWeight
(AxisConfig) Font weight of the title. This can be either a string (e.g "bold", "normal") or a number (100, 200, 300, ..., 900 where "normal" = 400 and "bold" = 700).

titleLimit
(AxisConfig) Maximum allowed pixel width of axis titles.

titleLineHeight
(AxisConfig) Line height in pixels for multi-line title text.

titleOpacity
(AxisConfig) Opacity of the axis title.

titlePadding
(AxisConfig) The padding, in pixels, between title and axis.

titleX
(AxisConfig) X-coordinate of the axis title relative to the axis group.

titleY
(AxisConfig) Y-coordinate of the axis title relative to the axis group.

translate
(AxisConfig) Translation offset in pixels applied to the axis group mark x and y. If specified, overrides the default behavior of a 0.5 offset to pixel-align stroked lines.

Value

A modified Vega-Lite Spec