vl_config_axisXTemporal

Add axisXTemporal config (AxisConfig) to a vega-lite spec.

vl_config_axisXTemporal(
  spec,
  .object = NULL,
  aria = NULL,
  bandPosition = NULL,
  description = NULL,
  disable = NULL,
  domain = NULL,
  domainCap = NULL,
  domainColor = NULL,
  domainDash = NULL,
  domainDashOffset = NULL,
  domainOpacity = NULL,
  domainWidth = NULL,
  format = NULL,
  formatType = NULL,
  grid = NULL,
  gridCap = NULL,
  gridColor = NULL,
  gridDash = NULL,
  gridDashOffset = NULL,
  gridOpacity = NULL,
  gridWidth = NULL,
  labelAlign = NULL,
  labelAngle = NULL,
  labelBaseline = NULL,
  labelBound = NULL,
  labelColor = NULL,
  labelExpr = NULL,
  labelFlush = NULL,
  labelFlushOffset = NULL,
  labelFont = NULL,
  labelFontSize = NULL,
  labelFontStyle = NULL,
  labelFontWeight = NULL,
  labelLimit = NULL,
  labelLineHeight = NULL,
  labelOffset = NULL,
  labelOpacity = NULL,
  labelOverlap = NULL,
  labelPadding = NULL,
  labelSeparation = NULL,
  labels = NULL,
  maxExtent = NULL,
  minExtent = NULL,
  offset = NULL,
  orient = NULL,
  position = NULL,
  style = NULL,
  tickBand = NULL,
  tickCap = NULL,
  tickColor = NULL,
  tickCount = NULL,
  tickDash = NULL,
  tickDashOffset = NULL,
  tickExtra = NULL,
  tickMinStep = NULL,
  tickOffset = NULL,
  tickOpacity = NULL,
  tickRound = NULL,
  tickSize = NULL,
  tickWidth = NULL,
  ticks = NULL,
  title = NULL,
  titleAlign = NULL,
  titleAnchor = NULL,
  titleAngle = NULL,
  titleBaseline = NULL,
  titleColor = NULL,
  titleFont = NULL,
  titleFontSize = NULL,
  titleFontStyle = NULL,
  titleFontWeight = NULL,
  titleLimit = NULL,
  titleLineHeight = NULL,
  titleOpacity = NULL,
  titlePadding = NULL,
  titleX = NULL,
  titleY = NULL,
  translate = NULL,
  values = NULL,
  zindex = NULL
)

Arguments

spec
An input vega-lite spec

.object
Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: AxisConfig

aria
A boolean flag indicating if ARIA attributes should be included (SVG output only). If false, the "aria-hidden" attribute will be set on the output SVG group, removing the axis from the ARIA accessibility tree.

Default value: true

Or: ExprRef

bandPosition
An interpolation fraction indicating where, for band scales, axis ticks should be positioned. A value of 0 places ticks at the left edge of their bands. A value of 0.5 places ticks in the middle of their bands.

Default value: 0.5

Or: ExprRef

description
A text description of this axis for ARIA accessibility (SVG output only). If the aria property is true, for SVG output the "aria-label" attribute will be set to this description. If the description is unspecified it will be automatically generated.

Or: ExprRef

disable
Disable axis by default.

domain
A boolean flag indicating if the domain (the axis baseline) should be included as part of the axis.

Default value: true

domainCap
The stroke cap for the domain line's ending style. One of "butt", "round" or "square".

Default value: "butt"

Or: ExprRef

domainColor
Color of axis domain line.

Default value: "gray".

Or: ExprRef

domainDash
An array of alternating [stroke, space] lengths for dashed domain lines.

Or: ExprRef

domainDashOffset
The pixel offset at which to start drawing with the domain dash array.

Or: ExprRef

domainOpacity
Opacity of the axis domain line.

Or: ExprRef

domainWidth
Stroke width of axis domain line

Default value: 1

Or: ExprRef

format
When used with the default "number" and "time" format type, the text formatting pattern for labels of guides (axes, legends, headers) and text marks.

If the format type is "number" (e.g., for quantitative fields), this is D3's number format pattern.

If the format type is "time" (e.g., for temporal fields), this is D3's time format pattern.

See the format documentation for more examples.

When used with a custom formatType, this value will be passed as format alongside datum.value to the registered function.

Default value: Derived from numberFormat config for number format and from timeFormat config for time format.

formatType
The format type for labels. One of "number", "time", or a registered custom format type.

Default value:

"time" for temporal fields and ordinal and nominal fields with timeUnit.

"number" for quantitative fields as well as ordinal and nominal fields without timeUnit.

grid
A boolean flag indicating if grid lines should be included as part of the axis

Default value: true for continuous scales that are not binned; otherwise, false.

gridCap
The stroke cap for grid lines' ending style. One of "butt", "round" or "square".

Default value: "butt"

Or: ExprRef

gridColor
Color of gridlines.

Default value: "lightGray".

Or: ExprRef

Or: ConditionalAxisColor

gridDash
An array of alternating [stroke, space] lengths for dashed grid lines.

Or: ExprRef

Or: ConditionalAxisNumberArray

gridDashOffset
The pixel offset at which to start drawing with the grid dash array.

Or: ExprRef

Or: ConditionalAxisNumber

gridOpacity
The stroke opacity of grid (value between [0,1])

Default value: 1

Or: ExprRef

Or: ConditionalAxisNumber

gridWidth
The grid width, in pixels.

Default value: 1

Or: ExprRef

Or: ConditionalAxisNumber

labelAlign
Horizontal text alignment of axis tick labels, overriding the default setting for the current axis orientation.

Or: ExprRef

Or: ConditionalAxisLabelAlign

labelAngle
The rotation angle of the axis labels.

Default value: -90 for nominal and ordinal fields; 0 otherwise.

Or: ExprRef

labelBaseline
Vertical text baseline of axis tick labels, overriding the default setting for the current axis orientation. One of "alphabetic" (default), "top", "middle", "bottom", "line-top", or "line-bottom". The "line-top" and "line-bottom" values operate similarly to "top" and "bottom", but are calculated relative to the lineHeight rather than fontSize alone.

Or: ExprRef

Or: ConditionalAxisLabelBaseline

labelBound
Indicates if labels should be hidden if they exceed the axis range. If false (the default) no bounds overlap analysis is performed. If true, labels will be hidden if they exceed the axis range by more than 1 pixel. If this property is a number, it specifies the pixel tolerance: the maximum amount by which a label bounding box may exceed the axis range.

Default value: false.

Or: ExprRef

labelColor
The color of the tick label, can be in hex color code or regular color name.

Or: ExprRef

Or: ConditionalAxisColor

labelExpr
Vega expression for customizing labels.

Note: The label text and value can be assessed via the label and value properties of the axis's backing datum object.

labelFlush
Indicates if the first and last axis labels should be aligned flush with the scale range. Flush alignment for a horizontal axis will left-align the first label and right-align the last label. For vertical axes, bottom and top text baselines are applied instead. If this property is a number, it also indicates the number of pixels by which to offset the first and last labels; for example, a value of 2 will flush-align the first and last labels and also push them 2 pixels outward from the center of the axis. The additional adjustment can sometimes help the labels better visually group with corresponding axis ticks.

Default value: true for axis of a continuous x-scale. Otherwise, false.

labelFlushOffset
Indicates the number of pixels by which to offset flush-adjusted labels. For example, a value of 2 will push flush-adjusted labels 2 pixels outward from the center of the axis. Offsets can help the labels better visually group with corresponding axis ticks.

Default value: 0.

Or: ExprRef

labelFont
The font of the tick label.

Or: ExprRef

Or: ConditionalAxisString

labelFontSize
The font size of the label, in pixels.

Or: ExprRef

Or: ConditionalAxisNumber

labelFontStyle
Font style of the title.

Or: ExprRef

Or: ConditionalAxisLabelFontStyle

labelFontWeight
Font weight of axis tick labels.

Or: ExprRef

Or: ConditionalAxisLabelFontWeight

labelLimit
Maximum allowed pixel width of axis tick labels.

Default value: 180

Or: ExprRef

labelLineHeight
Line height in pixels for multi-line label text or label text with "line-top" or "line-bottom" baseline.

Or: ExprRef

labelOffset
Position offset in pixels to apply to labels, in addition to tickOffset.

Default value: 0

Or: ExprRef

Or: ConditionalAxisNumber

labelOpacity
The opacity of the labels.

Or: ExprRef

Or: ConditionalAxisNumber

labelOverlap
The strategy to use for resolving overlap of axis labels. If false (the default), no overlap reduction is attempted. If set to true or "parity", a strategy of removing every other label is used (this works well for standard linear axes). If set to "greedy", a linear scan of the labels is performed, removing any labels that overlaps with the last visible label (this often works better for log-scaled axes).

Default value: true for non-nominal fields with non-log scales; "greedy" for log scales; otherwise false.

labelPadding
The padding in pixels between labels and ticks.

Default value: 2

Or: ExprRef

Or: ConditionalAxisNumber

labelSeparation
The minimum separation that must be between label bounding boxes for them to be considered non-overlapping (default 0). This property is ignored if labelOverlap resolution is not enabled.

Or: ExprRef

labels
A boolean flag indicating if labels should be included as part of the axis.

Default value: true.

maxExtent
The maximum extent in pixels that axis ticks and labels should use. This determines a maximum offset value for axis titles.

Default value: undefined.

Or: ExprRef

minExtent
The minimum extent in pixels that axis ticks and labels should use. This determines a minimum offset value for axis titles.

Default value: 30 for y-axis; undefined for x-axis.

Or: ExprRef

offset
The offset, in pixels, by which to displace the axis from the edge of the enclosing group or data rectangle.

Default value: derived from the axis config's offset (0 by default)

orient
The orientation of the axis. One of "top", "bottom", "left" or "right". The orientation can be used to further specialize the axis type (e.g., a y-axis oriented towards the right edge of the chart).

Default value: "bottom" for x-axes and "left" for y-axes.

position
The anchor position of the axis in pixels. For x-axes with top or bottom orientation, this sets the axis group x coordinate. For y-axes with left or right orientation, this sets the axis group y coordinate.

Default value: 0

style
A string or array of strings indicating the name of custom styles to apply to the axis. A style is a named collection of axis property defined within the style configuration. If style is an array, later styles will override earlier styles.

Default value: (none) Note: Any specified style will augment the default style. For example, an x-axis mark with "style": "foo" will use config.axisX and config.style.foo (the specified style "foo" has higher precedence).

tickBand
For band scales, indicates if ticks and grid lines should be placed at the "center" of a band (default) or at the band "extent"s to indicate intervals

Or: ExprRef

tickCap
The stroke cap for the tick lines' ending style. One of "butt", "round" or "square".

Default value: "butt"

Or: ExprRef

tickColor
The color of the axis's tick.

Default value: "gray"

Or: ExprRef

Or: ConditionalAxisColor

tickCount
A desired number of ticks, for axes visualizing quantitative scales. The resulting number may be different so that values are "nice" (multiples of 2, 5, 10) and lie within the underlying scale's range.

For scales of type "time" or "utc", the tick count can instead be a time interval specifier. Legal string values are "millisecond", "second", "minute", "hour", "day", "week", "month", and "year". Alternatively, an object-valued interval specifier of the form {"interval": "month", "step": 3} includes a desired number of interval steps. Here, ticks are generated for each quarter (Jan, Apr, Jul, Oct) boundary.

Default value: Determine using a formula ceil(width/40) for x and ceil(height/40) for y.

tickDash
An array of alternating [stroke, space] lengths for dashed tick mark lines.

Or: ExprRef

Or: ConditionalAxisNumberArray

tickDashOffset
The pixel offset at which to start drawing with the tick mark dash array.

Or: ExprRef

Or: ConditionalAxisNumber

tickExtra
Boolean flag indicating if an extra axis tick should be added for the initial position of the axis. This flag is useful for styling axes for band scales such that ticks are placed on band boundaries rather in the middle of a band. Use in conjunction with "bandPosition": 1 and an axis "padding" value of 0.

tickMinStep
The minimum desired step between axis ticks, in terms of scale domain values. For example, a value of 1 indicates that ticks should not be less than 1 unit apart. If tickMinStep is specified, the tickCount value will be adjusted, if necessary, to enforce the minimum step value.

tickOffset
Position offset in pixels to apply to ticks, labels, and gridlines.

Or: ExprRef

tickOpacity
Opacity of the ticks.

Or: ExprRef

Or: ConditionalAxisNumber

tickRound
Boolean flag indicating if pixel position values should be rounded to the nearest integer.

Default value: true

tickSize
The size in pixels of axis ticks.

Default value: 5

Or: ExprRef

Or: ConditionalAxisNumber

tickWidth
The width, in pixels, of ticks.

Default value: 1

Or: ExprRef

Or: ConditionalAxisNumber

ticks
Boolean value that determines whether the axis should include ticks.

Default value: true

title
A title for the field. If null, the title will be removed.

Default value: derived from the field's name and transformation function (aggregate, bin and timeUnit). If the field has an aggregate function, the function is displayed as part of the title (e.g., "Sum of Profit"). If the field is binned or has a time unit applied, the applied function is shown in parentheses (e.g., "Profit (binned)", "Transaction Date (year-month)"). Otherwise, the title is simply the field name.

Notes:

You can customize the default field title format by providing the fieldTitle property in the config or fieldTitle function via the compile function's options.

If both field definition's title and axis, header, or legend title are defined, axis/header/legend title will be used.

titleAlign
Horizontal text alignment of axis titles.

Or: ExprRef

titleAnchor
Text anchor position for placing axis titles.

Or: ExprRef

titleAngle
Angle in degrees of axis titles.

Or: ExprRef

titleBaseline
Vertical text baseline for axis titles. One of "alphabetic" (default), "top", "middle", "bottom", "line-top", or "line-bottom". The "line-top" and "line-bottom" values operate similarly to "top" and "bottom", but are calculated relative to the lineHeight rather than fontSize alone.

Or: ExprRef

titleColor
Color of the title, can be in hex color code or regular color name.

Or: ExprRef

titleFont
Font of the title. (e.g., "Helvetica Neue").

Or: ExprRef

titleFontSize
Font size of the title.

Or: ExprRef

titleFontStyle
Font style of the title.

Or: ExprRef

titleFontWeight
Font weight of the title. This can be either a string (e.g "bold", "normal") or a number (100, 200, 300, ..., 900 where "normal" = 400 and "bold" = 700).

Or: ExprRef

titleLimit
Maximum allowed pixel width of axis titles.

Or: ExprRef

titleLineHeight
Line height in pixels for multi-line title text or title text with "line-top" or "line-bottom" baseline.

Or: ExprRef

titleOpacity
Opacity of the axis title.

Or: ExprRef

titlePadding
The padding, in pixels, between title and axis.

Or: ExprRef

titleX
X-coordinate of the axis title relative to the axis group.

Or: ExprRef

titleY
Y-coordinate of the axis title relative to the axis group.

Or: ExprRef

translate
Coordinate space translation offset for axis layout. By default, axes are translated by a 0.5 pixel offset for both the x and y coordinates in order to align stroked lines with the pixel grid. However, for vector graphics output these pixel-specific adjustments may be undesirable, in which case translate can be changed (for example, to zero).

Default value: 0.5

Or: ExprRef

values
Explicitly set the visible axis tick values.

zindex
A non-negative integer indicating the z-index of the axis. If zindex is 0, axes should be drawn behind all chart elements. To put them in front, set zindex to 1 or more.

Default value: 0 (behind the marks).

Value

A modified Vega-Lite Spec