vl_config_legend

Add legend config (LegendConfig) to a vega-lite spec.

Usage

vl_config_legend(
  spec,
  aria = NULL,
  clipHeight = NULL,
  columnPadding = NULL,
  columns = NULL,
  cornerRadius = NULL,
  description = NULL,
  direction = NULL,
  disable = NULL,
  fillColor = NULL,
  gradientDirection = NULL,
  gradientHorizontalMaxLength = NULL,
  gradientHorizontalMinLength = NULL,
  gradientLabelLimit = NULL,
  gradientLabelOffset = NULL,
  gradientLength = NULL,
  gradientOpacity = NULL,
  gradientStrokeColor = NULL,
  gradientStrokeWidth = NULL,
  gradientThickness = NULL,
  gradientVerticalMaxLength = NULL,
  gradientVerticalMinLength = NULL,
  gridAlign = NULL,
  labelAlign = NULL,
  labelBaseline = NULL,
  labelColor = NULL,
  labelFont = NULL,
  labelFontSize = NULL,
  labelFontStyle = NULL,
  labelFontWeight = NULL,
  labelLimit = NULL,
  labelOffset = NULL,
  labelOpacity = NULL,
  labelOverlap = NULL,
  labelPadding = NULL,
  labelSeparation = NULL,
  layout = NULL,
  legendX = NULL,
  legendY = NULL,
  offset = NULL,
  orient = NULL,
  padding = NULL,
  rowPadding = NULL,
  strokeColor = NULL,
  strokeDash = NULL,
  strokeWidth = NULL,
  symbolBaseFillColor = NULL,
  symbolBaseStrokeColor = NULL,
  symbolDash = NULL,
  symbolDashOffset = NULL,
  symbolDirection = NULL,
  symbolFillColor = NULL,
  symbolLimit = NULL,
  symbolOffset = NULL,
  symbolOpacity = NULL,
  symbolSize = NULL,
  symbolStrokeColor = NULL,
  symbolStrokeWidth = NULL,
  symbolType = NULL,
  tickCount = NULL,
  title = NULL,
  titleAlign = NULL,
  titleAnchor = NULL,
  titleBaseline = NULL,
  titleColor = NULL,
  titleFont = NULL,
  titleFontSize = NULL,
  titleFontStyle = NULL,
  titleFontWeight = NULL,
  titleLimit = NULL,
  titleLineHeight = NULL,
  titleOpacity = NULL,
  titleOrient = NULL,
  titlePadding = NULL,
  unselectedOpacity = NULL,
  zindex = NULL,
  .object = NULL
)

Arguments

spec

An input vega-lite spec

aria

A boolean flag indicating if ARIA attributes should be included (SVG output only). If false, the "aria-hidden" attribute will be set on the output SVG group, removing the legend from the ARIA accessibility tree.

Default value: true

Or: ExprRef

clipHeight

The height in pixels to clip symbol legend entries and limit their size.

Or: ExprRef

columnPadding

The horizontal padding in pixels between symbol legend entries.

Default value: 10.

Or: ExprRef

columns

The number of columns in which to arrange symbol legend entries. A value of 0 or lower indicates a single row with one column per entry.

Or: ExprRef

cornerRadius

Corner radius for the full legend.

Or: ExprRef

description

A text description of this legend for ARIA accessibility (SVG output only). If the aria property is true, for SVG output the "aria-label" attribute will be set to this description. If the description is unspecified it will be automatically generated.

Or: ExprRef

direction

The direction of the legend, one of "vertical" or "horizontal".

Default value:

For top-/bottom-oriented legends, "horizontal"

For left-/right-oriented legends, "vertical"

For top/bottom-left/right-oriented legends, "horizontal" for gradient legends and "vertical" for symbol legends.

disable

Disable legend by default

fillColor

Background fill color for the full legend.

Or: ExprRef

gradientDirection

The default direction ("horizontal" or "vertical") for gradient legends.

Default value: "vertical".

Or: ExprRef

gradientHorizontalMaxLength

Max legend length for a horizontal gradient when config.legend.gradientLength is undefined.

Default value: 200

gradientHorizontalMinLength

Min legend length for a horizontal gradient when config.legend.gradientLength is undefined.

Default value: 100

gradientLabelLimit

The maximum allowed length in pixels of color ramp gradient labels.

Or: ExprRef

gradientLabelOffset

Vertical offset in pixels for color ramp gradient labels.

Default value: 2.

Or: ExprRef

gradientLength

The length in pixels of the primary axis of a color gradient. This value corresponds to the height of a vertical gradient or the width of a horizontal gradient.

Default value: 200.

Or: ExprRef

gradientOpacity

Opacity of the color gradient.

Or: ExprRef

gradientStrokeColor

The color of the gradient stroke, can be in hex color code or regular color name.

Default value: "lightGray".

Or: ExprRef

gradientStrokeWidth

The width of the gradient stroke, in pixels.

Default value: 0.

Or: ExprRef

gradientThickness

The thickness in pixels of the color gradient. This value corresponds to the width of a vertical gradient or the height of a horizontal gradient.

Default value: 16.

Or: ExprRef

gradientVerticalMaxLength

Max legend length for a vertical gradient when config.legend.gradientLength is undefined.

Default value: 200

gradientVerticalMinLength

Min legend length for a vertical gradient when config.legend.gradientLength is undefined.

Default value: 100

gridAlign

The alignment to apply to symbol legends rows and columns. The supported string values are "all", "each" (the default), and none. For more information, see the grid layout documentation.

Default value: "each".

Or: ExprRef

labelAlign

The alignment of the legend label, can be left, center, or right.

Or: ExprRef

labelBaseline

The position of the baseline of legend label, can be "top", "middle", "bottom", or "alphabetic".

Default value: "middle".

Or: ExprRef

labelColor

The color of the legend label, can be in hex color code or regular color name.

Or: ExprRef

labelFont

The font of the legend label.

Or: ExprRef

labelFontSize

The font size of legend label.

Default value: 10.

Or: ExprRef

labelFontStyle

The font style of legend label.

Or: ExprRef

labelFontWeight

The font weight of legend label.

Or: ExprRef

labelLimit

Maximum allowed pixel width of legend tick labels.

Default value: 160.

Or: ExprRef

labelOffset

The offset of the legend label.

Default value: 4.

Or: ExprRef

labelOpacity

Opacity of labels.

Or: ExprRef

labelOverlap

The strategy to use for resolving overlap of labels in gradient legends. If false, no overlap reduction is attempted. If set to true or "parity", a strategy of removing every other label is used. If set to "greedy", a linear scan of the labels is performed, removing any label that overlaps with the last visible label (this often works better for log-scaled axes).

Default value: "greedy" for log scales otherwise true`.

labelPadding

Padding in pixels between the legend and legend labels.

Or: ExprRef

labelSeparation

The minimum separation that must be between label bounding boxes for them to be considered non-overlapping (default 0). This property is ignored if labelOverlap resolution is not enabled.

Or: ExprRef

layout

ExprRef

legendX

Custom x-position for legend with orient "none".

Or: ExprRef

legendY

Custom y-position for legend with orient "none".

Or: ExprRef

offset

The offset in pixels by which to displace the legend from the data rectangle and axes.

Default value: 18.

Or: ExprRef

orient

The orientation of the legend, which determines how the legend is positioned within the scene. One of "left", "right", "top", "bottom", "top-left", "top-right", "bottom-left", "bottom-right", "none".

Default value: "right"

padding

The padding between the border and content of the legend group.

Default value: 0.

Or: ExprRef

rowPadding

The vertical padding in pixels between symbol legend entries.

Default value: 2.

Or: ExprRef

strokeColor

Border stroke color for the full legend.

Or: ExprRef

strokeDash

Border stroke dash pattern for the full legend.

Or: ExprRef

strokeWidth

Border stroke width for the full legend.

Or: ExprRef

symbolBaseFillColor

Default fill color for legend symbols. Only applied if there is no "fill" scale color encoding for the legend.

Default value: "transparent".

Or: ExprRef

symbolBaseStrokeColor

Default stroke color for legend symbols. Only applied if there is no "fill" scale color encoding for the legend.

Default value: "gray".

Or: ExprRef

symbolDash

An array of alternating [stroke, space] lengths for dashed symbol strokes.

Or: ExprRef

symbolDashOffset

The pixel offset at which to start drawing with the symbol stroke dash array.

Or: ExprRef

symbolDirection

The default direction ("horizontal" or "vertical") for symbol legends.

Default value: "vertical".

Or: ExprRef

symbolFillColor

The color of the legend symbol,

Or: ExprRef

symbolLimit

The maximum number of allowed entries for a symbol legend. Additional entries will be dropped.

Or: ExprRef

symbolOffset

Horizontal pixel offset for legend symbols.

Default value: 0.

Or: ExprRef

symbolOpacity

Opacity of the legend symbols.

Or: ExprRef

symbolSize

The size of the legend symbol, in pixels.

Default value: 100.

Or: ExprRef

symbolStrokeColor

Stroke color for legend symbols.

Or: ExprRef

symbolStrokeWidth

The width of the symbol's stroke.

Default value: 1.5.

Or: ExprRef

symbolType

The symbol shape. One of the plotting shapes circle (default), square, cross, diamond, triangle-up, triangle-down, triangle-right, or triangle-left, the line symbol stroke, or one of the centered directional shapes arrow, wedge, or triangle. Alternatively, a custom SVG path string can be provided. For correct sizing, custom shape paths should be defined within a square bounding box with coordinates ranging from -1 to 1 along both the x and y dimensions.

Default value: "circle".

Or: ExprRef

tickCount

The desired number of tick values for quantitative legends.

Or: ExprRef

title

Set to null to disable title for the axis, legend, or header.

titleAlign

Horizontal text alignment for legend titles.

Default value: "left".

Or: ExprRef

titleAnchor

Text anchor position for placing legend titles.

Or: ExprRef

titleBaseline

Vertical text baseline for legend titles. One of "alphabetic" (default), "top", "middle", "bottom", "line-top", or "line-bottom". The "line-top" and "line-bottom" values operate similarly to "top" and "bottom", but are calculated relative to the lineHeight rather than fontSize alone.

Default value: "top".

Or: ExprRef

titleColor

The color of the legend title, can be in hex color code or regular color name.

Or: ExprRef

titleFont

The font of the legend title.

Or: ExprRef

titleFontSize

The font size of the legend title.

Or: ExprRef

titleFontStyle

The font style of the legend title.

Or: ExprRef

titleFontWeight

The font weight of the legend title. This can be either a string (e.g "bold", "normal") or a number (100, 200, 300, ..., 900 where "normal" = 400 and "bold" = 700).

Or: ExprRef

titleLimit

Maximum allowed pixel width of legend titles.

Default value: 180.

Or: ExprRef

titleLineHeight

Line height in pixels for multi-line title text or title text with "line-top" or "line-bottom" baseline.

Or: ExprRef

titleOpacity

Opacity of the legend title.

Or: ExprRef

titleOrient

Orientation of the legend title.

Or: ExprRef

titlePadding

The padding, in pixels, between title and legend.

Default value: 5.

Or: ExprRef

unselectedOpacity

The opacity of unselected legend entries.

Default value: 0.35.

zindex

The integer z-index indicating the layering of the legend group relative to other axis, mark, and legend groups.

Or: ExprRef

.object

Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than spec. Objects can be of type: LegendConfig

Value

A modified Vega-Lite Spec