vl_config_projection

Add projection config (ProjectionConfig) to a vega-lite spec.

vl_config_projection(
  spec,
  .object = NULL,
  center = NULL,
  clipAngle = NULL,
  clipExtent = NULL,
  coefficient = NULL,
  distance = NULL,
  extent = NULL,
  fit = NULL,
  fraction = NULL,
  lobes = NULL,
  parallel = NULL,
  parallels = NULL,
  pointRadius = NULL,
  precision = NULL,
  radius = NULL,
  ratio = NULL,
  reflectX = NULL,
  reflectY = NULL,
  rotate = NULL,
  scale = NULL,
  size = NULL,
  spacing = NULL,
  tilt = NULL,
  translate = NULL,
  type = NULL
)

Arguments

spec
An input vega-lite spec

.object
Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: Projection

center
The projection's center, a two-element array of longitude and latitude in degrees.

Default value: [0, 0]

Or: ExprRef

clipAngle
The projection's clipping circle radius to the specified angle in degrees. If null, switches to antimeridian cutting rather than small-circle clipping.

Or: ExprRef

clipExtent
The projection's viewport clip extent to the specified bounds in pixels. The extent bounds are specified as an array [\[x0, y0\], \[x1, y1\]], where x0 is the left-side of the viewport, y0 is the top, x1 is the right and y1 is the bottom. If null, no viewport clipping is performed.

Or: ExprRef

coefficient
The coefficient parameter for the hammer projection.

Default value: 2

Or: ExprRef

distance
For the satellite projection, the distance from the center of the sphere to the point of view, as a proportion of the sphere’s radius. The recommended maximum clip angle for a given distance is acos(1 / distance) converted to degrees. If tilt is also applied, then more conservative clipping may be necessary.

Default value: 2.0

Or: ExprRef

extent
Vector2<Vector2<number>>

Or: ExprRef

fit
array

Or: ExprRef

fraction
The fraction parameter for the bottomley projection.

Default value: 0.5, corresponding to a sin(ψ) where ψ = π/6.

Or: ExprRef

lobes
The number of lobes in projections that support multi-lobe views: berghaus, gingery, or healpix. The default value varies based on the projection type.

Or: ExprRef

parallel
The parallel parameter for projections that support it: armadillo, bonne, craig, cylindricalEqualArea, cylindricalStereographic, hammerRetroazimuthal, loximuthal, or rectangularPolyconic. The default value varies based on the projection type.

Or: ExprRef

parallels
For conic projections, the two standard parallels that define the map layout. The default depends on the specific conic projection used.

Or: ExprRef

pointRadius
The default radius (in pixels) to use when drawing GeoJSON Point and MultiPoint geometries. This parameter sets a constant default value. To modify the point radius in response to data, see the corresponding parameter of the GeoPath and GeoShape transforms.

Default value: 4.5

Or: ExprRef

precision
The threshold for the projection's adaptive resampling to the specified value in pixels. This value corresponds to the Douglas–Peucker distance. If precision is not specified, returns the projection's current resampling precision which defaults to √0.5 ≅ 0.70710….

Or: ExprRef

radius
The radius parameter for the airy or gingery projection. The default value varies based on the projection type.

Or: ExprRef

ratio
The ratio parameter for the hill, hufnagel, or wagner projections. The default value varies based on the projection type.

Or: ExprRef

reflectX
Sets whether or not the x-dimension is reflected (negated) in the output.

Or: ExprRef

reflectY
Sets whether or not the y-dimension is reflected (negated) in the output.

Or: ExprRef

rotate
The projection's three-axis rotation to the specified angles, which must be a two- or three-element array of numbers [lambda, phi, gamma] specifying the rotation angles in degrees about each spherical axis. (These correspond to yaw, pitch and roll.)

Default value: [0, 0, 0]

Or: ExprRef

scale
The projection’s scale (zoom) factor, overriding automatic fitting. The default scale is projection-specific. The scale factor corresponds linearly to the distance between projected points; however, scale factor values are not equivalent across projections.

size
Used in conjunction with fit, provides the width and height in pixels of the area to which the projection should be automatically fit.

Or: ExprRef

spacing
The spacing parameter for the lagrange projection.

Default value: 0.5

Or: ExprRef

tilt
The tilt angle (in degrees) for the satellite projection.

Default value: 0.

Or: ExprRef

translate
The projection’s translation offset as a two-element array [tx, ty].

type
The cartographic projection to use. This value is case-insensitive, for example "albers" and "Albers" indicate the same projection type. You can find all valid projection types in the documentation.

Default value: equalEarth

Value

A modified Vega-Lite Spec