vl_config_rule

Add rule config (MarkConfig) to a vega-lite spec.

vl_config_rule(spec, .object = NULL, align = NULL, angle = NULL,
  aspect = NULL, baseline = NULL, color = NULL,
  cornerRadius = NULL, cornerRadiusBottomLeft = NULL,
  cornerRadiusBottomRight = NULL, cornerRadiusTopLeft = NULL,
  cornerRadiusTopRight = NULL, cursor = NULL, dir = NULL,
  dx = NULL, dy = NULL, ellipsis = NULL, fill = NULL,
  fillOpacity = NULL, filled = NULL, font = NULL, fontSize = NULL,
  fontStyle = NULL, fontWeight = NULL, height = NULL, href = NULL,
  interpolate = NULL, invalid = NULL, limit = NULL,
  lineBreak = NULL, lineHeight = NULL, opacity = NULL,
  order = NULL, orient = NULL, radius = NULL, shape = NULL,
  size = NULL, stroke = NULL, strokeCap = NULL, strokeDash = NULL,
  strokeDashOffset = NULL, strokeJoin = NULL,
  strokeMiterLimit = NULL, strokeOpacity = NULL, strokeWidth = NULL,
  tension = NULL, text = NULL, theta = NULL, timeUnitBand = NULL,
  timeUnitBandPosition = NULL, tooltip = NULL, width = NULL,
  x = NULL, x2 = NULL, y = NULL, y2 = NULL)

Arguments

spec
An input vega-lite spec

.object
Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: MarkConfig

align
(MarkConfig) The horizontal alignment of the text or ranged marks (area, bar, image, rect, rule). One of "left", "right", "center".

angle
(MarkConfig) The rotation angle of the text, in degrees.

aspect
(MarkConfig) Whether to keep aspect ratio of image marks.

baseline
(MarkConfig) The vertical alignment of the text or ranged marks (area, bar, image, rect, rule). One of "top", "middle", "bottom".

Default value: "middle"

color
(MarkConfig) Default color.

Default value: <span style="color: #4682b4;">■</span> "#4682b4"

Note:

This property cannot be used in a style config.

The fill and stroke properties have higher precedence than color and will override color.

cornerRadius
(MarkConfig) The radius in pixels of rounded rectangle corners.

Default value: 0

cornerRadiusBottomLeft
(MarkConfig) The radius in pixels of rounded rectangle bottom left corner.

Default value: 0

cornerRadiusBottomRight
(MarkConfig) The radius in pixels of rounded rectangle bottom right corner.

Default value: 0

cornerRadiusTopLeft
(MarkConfig) The radius in pixels of rounded rectangle top right corner.

Default value: 0

cornerRadiusTopRight
(MarkConfig) The radius in pixels of rounded rectangle top left corner.

Default value: 0

cursor
(MarkConfig) The mouse cursor used over the mark. Any valid CSS cursor type can be used.

dir
(MarkConfig) The direction of the text. One of "ltr" (left-to-right) or "rtl" (right-to-left). This property determines on which side is truncated in response to the limit parameter.

Default value: "ltr"

dx
(MarkConfig) The horizontal offset, in pixels, between the text label and its anchor point. The offset is applied after rotation by the angle property.

dy
(MarkConfig) The vertical offset, in pixels, between the text label and its anchor point. The offset is applied after rotation by the angle property.

ellipsis
(MarkConfig) The ellipsis string for text truncated in response to the limit parameter.

Default value: "…"

fill
(MarkConfig) Default Fill Color. This has higher precedence than config.color.

Default value: (None)

fillOpacity
(MarkConfig) The fill opacity (value between [0,1]).

Default value: 1

filled
(MarkConfig) Whether the mark's color should be used as fill color instead of stroke color.

Default value: false for all point, line, and rule marks as well as geoshape marks for graticule data sources; otherwise, true.

Note: This property cannot be used in a style config.

font
(MarkConfig) The typeface to set the text in (e.g., "Helvetica Neue").

fontSize
(MarkConfig) The font size, in pixels.

fontStyle
(MarkConfig) The font style (e.g., "italic").

fontWeight
(MarkConfig) The font weight. This can be either a string (e.g "bold", "normal") or a number (100, 200, 300, ..., 900 where "normal" = 400 and "bold" = 700).

height
(MarkConfig) Height of the marks.

href
(MarkConfig) A URL to load upon mouse click. If defined, the mark acts as a hyperlink.

interpolate
(MarkConfig) The line interpolation method to use for line and area marks. One of the following:

"linear": piecewise linear segments, as in a polyline.

"linear-closed": close the linear segments to form a polygon.

"step": alternate between horizontal and vertical segments, as in a step function.

"step-before": alternate between vertical and horizontal segments, as in a step function.

"step-after": alternate between horizontal and vertical segments, as in a step function.

"basis": a B-spline, with control point duplication on the ends.

"basis-open": an open B-spline; may not intersect the start or end.

"basis-closed": a closed B-spline, as in a loop.

"cardinal": a Cardinal spline, with control point duplication on the ends.

"cardinal-open": an open Cardinal spline; may not intersect the start or end, but will intersect other control points.

"cardinal-closed": a closed Cardinal spline, as in a loop.

"bundle": equivalent to basis, except the tension parameter is used to straighten the spline.

"monotone": cubic interpolation that preserves monotonicity in y.

invalid
(MarkConfig) Defines how Vega-Lite should handle marks for invalid values (null and NaN).

If set to "filter" (default), all data items with null values will be skipped (for line, trail, and area marks) or filtered (for other marks).

If null, all data items are included. In this case, invalid values will be interpreted as zeroes.

limit
(MarkConfig) The maximum length of the text mark in pixels. The text value will be automatically truncated if the rendered size exceeds the limit.

Default value: 0, indicating no limit

lineBreak
(MarkConfig) A delimiter, such as a newline character, upon which to break text strings into multiple lines. This property will be ignored if the text property is array-valued.

lineHeight
(MarkConfig) The height, in pixels, of each line of text in a multi-line text mark.

opacity
(MarkConfig) The overall opacity (value between [0,1]).

Default value: 0.7 for non-aggregate plots with point, tick, circle, or square marks or layered bar charts and 1 otherwise.

order
(MarkConfig) For line and trail marks, this order property can be set to null or false to make the lines use the original order in the data sources.

orient
(MarkConfig) The orientation of a non-stacked bar, tick, area, and line charts. The value is either horizontal (default) or vertical.

For bar, rule and tick, this determines whether the size of the bar and tick should be applied to x or y dimension.

For area, this property determines the orient property of the Vega output.

For line and trail marks, this property determines the sort order of the points in the line if config.sortLineBy is not specified. For stacked charts, this is always determined by the orientation of the stack; therefore explicitly specified value will be ignored.

radius
(MarkConfig) Polar coordinate radial offset, in pixels, of the text label from the origin determined by the x and y properties.

shape
(MarkConfig) Shape of the point marks. Supported values include:

plotting shapes: "circle", "square", "cross", "diamond", "triangle-up", "triangle-down", "triangle-right", or "triangle-left".

the line symbol "stroke"

centered directional shapes "arrow", "wedge", or "triangle"

a custom SVG path string (For correct sizing, custom shape paths should be defined within a square bounding box with coordinates ranging from -1 to 1 along both the x and y dimensions.)

Default value: "circle"

size
(MarkConfig) Default size for marks.

For point/circle/square, this represents the pixel area of the marks. For example: in the case of circles, the radius is determined in part by the square root of the size value.

For bar, this represents the band size of the bar, in pixels.

For text, this represents the font size, in pixels.

Default value:

30 for point, circle, square marks; width/height's step

2 for bar marks with discrete dimensions;

5 for bar marks with continuous dimensions;

11 for text marks.

stroke
(MarkConfig) Default Stroke Color. This has higher precedence than config.color.

Default value: (None)

strokeCap
(MarkConfig) The stroke cap for line ending style. One of "butt", "round", or "square".

Default value: "square"

strokeDash
(MarkConfig) An array of alternating stroke, space lengths for creating dashed or dotted lines.

strokeDashOffset
(MarkConfig) The offset (in pixels) into which to begin drawing with the stroke dash array.

strokeJoin
(MarkConfig) The stroke line join method. One of "miter", "round" or "bevel".

Default value: "miter"

strokeMiterLimit
(MarkConfig) The miter limit at which to bevel a line join.

strokeOpacity
(MarkConfig) The stroke opacity (value between [0,1]).

Default value: 1

strokeWidth
(MarkConfig) The stroke width, in pixels.

tension
(MarkConfig) Depending on the interpolation type, sets the tension parameter (for line and area marks).

text
(MarkConfig) Placeholder text if the text channel is not specified

theta
(MarkConfig) Polar coordinate angle, in radians, of the text label from the origin determined by the x and y properties. Values for theta follow the same convention of arc mark startAngle and endAngle properties: angles are measured in radians, with 0 indicating "north".

timeUnitBand
(MarkConfig) Default relative band size for a time unit. If set to 1, the bandwidth of the marks will be equal to the time unit band step. If set to 0.5, bandwidth of the marks will be half of the time unit band step.

timeUnitBandPosition
(MarkConfig) Default relative band position for a time unit. If set to 0, the marks will be positioned at the beginning of the time unit band step. If set to 0.5, the marks will be positioned in the middle of the time unit band step.

tooltip
(MarkConfig) The tooltip text string to show upon mouse hover or an object defining which fields should the tooltip be derived from.

If tooltip is true or {"content": "encoding"}, then all fields from encoding will be used.

If tooltip is {"content": "data"}, then all fields that appear in the highlighted data point will be used.

If set to null or false, then no tooltip will be used.

See the tooltip documentation for a detailed discussion about tooltip in Vega-Lite.

Default value: null

width
(MarkConfig) Width of the marks.

x
(MarkConfig) X coordinates of the marks, or width of horizontal "bar" and "area" without specified x2 or width.

The value of this channel can be a number or a string "width" for the width of the plot.

x2
(MarkConfig) X2 coordinates for ranged "area", "bar", "rect", and "rule".

The value of this channel can be a number or a string "width" for the width of the plot.

y
(MarkConfig) Y coordinates of the marks, or height of vertical "bar" and "area" without specified y2 or height.

The value of this channel can be a number or a string "height" for the height of the plot.

y2
(MarkConfig) Y2 coordinates for ranged "area", "bar", "rect", and "rule".

The value of this channel can be a number or a string "height" for the height of the plot.

Value

A modified Vega-Lite Spec