vl_config_scale

Add scale config (ScaleConfig) to a vega-lite spec.

vl_config_scale(
  spec,
  .object = NULL,
  bandPaddingInner = NULL,
  bandPaddingOuter = NULL,
  barBandPaddingInner = NULL,
  clamp = NULL,
  continuousPadding = NULL,
  maxBandSize = NULL,
  maxFontSize = NULL,
  maxOpacity = NULL,
  maxSize = NULL,
  maxStrokeWidth = NULL,
  minBandSize = NULL,
  minFontSize = NULL,
  minOpacity = NULL,
  minSize = NULL,
  minStrokeWidth = NULL,
  pointPadding = NULL,
  quantileCount = NULL,
  quantizeCount = NULL,
  rectBandPaddingInner = NULL,
  round = NULL,
  useUnaggregatedDomain = NULL,
  xReverse = NULL
)

Arguments

spec
An input vega-lite spec

.object
Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: ScaleConfig

bandPaddingInner
Default inner padding for x and y band-ordinal scales.

Default value:

barBandPaddingInner for bar marks (0.1 by default)

rectBandPaddingInner for rect and other marks (0 by default)

bandPaddingOuter
Default outer padding for x and y band-ordinal scales.

Default value: paddingInner/2 (which makes width/height = number of unique values * step)

barBandPaddingInner
Default inner padding for x and y band-ordinal scales of "bar" marks.

Default value: 0.1

clamp
If true, values that exceed the data domain are clamped to either the minimum or maximum range value

continuousPadding
Default padding for continuous scales.

Default: 5 for continuous x-scale of a vertical bar and continuous y-scale of a horizontal bar.; 0 otherwise.

maxBandSize
The default max value for mapping quantitative fields to bar's size/bandSize.

If undefined (default), we will use the axis's size (width or height) - 1.

maxFontSize
The default max value for mapping quantitative fields to text's size/fontSize.

Default value: 40

maxOpacity
Default max opacity for mapping a field to opacity.

Default value: 0.8

maxSize
Default max value for point size scale.

maxStrokeWidth
Default max strokeWidth for the scale of strokeWidth for rule and line marks and of size for trail marks.

Default value: 4

minBandSize
The default min value for mapping quantitative fields to bar and tick's size/bandSize scale with zero=false.

Default value: 2

minFontSize
The default min value for mapping quantitative fields to tick's size/fontSize scale with zero=false

Default value: 8

minOpacity
Default minimum opacity for mapping a field to opacity.

Default value: 0.3

minSize
Default minimum value for point size scale with zero=false.

Default value: 9

minStrokeWidth
Default minimum strokeWidth for the scale of strokeWidth for rule and line marks and of size for trail marks with zero=false.

Default value: 1

pointPadding
Default outer padding for x and y point-ordinal scales.

Default value: 0.5 (which makes width/height = number of unique values * step)

quantileCount
Default range cardinality for quantile scale.

Default value: 4

quantizeCount
Default range cardinality for quantize scale.

Default value: 4

rectBandPaddingInner
Default inner padding for x and y band-ordinal scales of "rect" marks.

Default value: 0

round
If true, rounds numeric output values to integers. This can be helpful for snapping to the pixel grid. (Only available for x, y, and size scales.)

useUnaggregatedDomain
Use the source data range before aggregation as scale domain instead of aggregated data for aggregate axis.

This is equivalent to setting domain to "unaggregate" for aggregated quantitative fields by default.

This property only works with aggregate functions that produce values within the raw data domain ("mean", "average", "median", "q1", "q3", "min", "max"). For other aggregations that produce values outside of the raw data domain (e.g. "count", "sum"), this property is ignored.

Default value: false

xReverse
Reverse x-scale by default (useful for right-to-left charts).

Value

A modified Vega-Lite Spec