vl_config_title

Add title config (TitleConfig) to a vega-lite spec.

vl_config_title(
  spec,
  .object = NULL,
  align = NULL,
  anchor = NULL,
  angle = NULL,
  aria = NULL,
  baseline = NULL,
  color = NULL,
  dx = NULL,
  dy = NULL,
  font = NULL,
  fontSize = NULL,
  fontStyle = NULL,
  fontWeight = NULL,
  frame = NULL,
  limit = NULL,
  lineHeight = NULL,
  offset = NULL,
  orient = NULL,
  subtitleColor = NULL,
  subtitleFont = NULL,
  subtitleFontSize = NULL,
  subtitleFontStyle = NULL,
  subtitleFontWeight = NULL,
  subtitleLineHeight = NULL,
  subtitlePadding = NULL,
  zindex = NULL
)

Arguments

spec
An input vega-lite spec

.object
Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: BaseTitleNoValueRefs

align
Horizontal text alignment for title text. One of "left", "center", or "right".

anchor
The anchor position for placing the title and subtitle text. One of "start", "middle", or "end". For example, with an orientation of top these anchor positions map to a left-, center-, or right-aligned title.

Or: ExprRef

angle
Angle in degrees of title and subtitle text.

Or: ExprRef

aria
A boolean flag indicating if ARIA attributes should be included (SVG output only). If false, the "aria-hidden" attribute will be set on the output SVG group, removing the title from the ARIA accessibility tree.

Default value: true

Or: ExprRef

baseline
Vertical text baseline for title and subtitle text. One of "alphabetic" (default), "top", "middle", "bottom", "line-top", or "line-bottom". The "line-top" and "line-bottom" values operate similarly to "top" and "bottom", but are calculated relative to the lineHeight rather than fontSize alone.

color
Text color for title text.

Or: ExprRef

dx
Delta offset for title and subtitle text x-coordinate.

Or: ExprRef

dy
Delta offset for title and subtitle text y-coordinate.

Or: ExprRef

font
Font name for title text.

Or: ExprRef

fontSize
Font size in pixels for title text.

Or: ExprRef

fontStyle
Font style for title text.

Or: ExprRef

fontWeight
Font weight for title text. This can be either a string (e.g "bold", "normal") or a number (100, 200, 300, ..., 900 where "normal" = 400 and "bold" = 700).

Or: ExprRef

frame
The reference frame for the anchor position, one of "bounds" (to anchor relative to the full bounding box) or "group" (to anchor relative to the group width or height).

Or: ExprRef

limit
The maximum allowed length in pixels of title and subtitle text.

Or: ExprRef

lineHeight
Line height in pixels for multi-line title text or title text with "line-top" or "line-bottom" baseline.

Or: ExprRef

offset
The orthogonal offset in pixels by which to displace the title group from its position along the edge of the chart.

Or: ExprRef

orient
Default title orientation ("top", "bottom", "left", or "right")

Or: ExprRef

subtitleColor
Text color for subtitle text.

Or: ExprRef

subtitleFont
Font name for subtitle text.

Or: ExprRef

subtitleFontSize
Font size in pixels for subtitle text.

Or: ExprRef

subtitleFontStyle
Font style for subtitle text.

Or: ExprRef

subtitleFontWeight
Font weight for subtitle text. This can be either a string (e.g "bold", "normal") or a number (100, 200, 300, ..., 900 where "normal" = 400 and "bold" = 700).

Or: ExprRef

subtitleLineHeight
Line height in pixels for multi-line subtitle text.

Or: ExprRef

subtitlePadding
The padding in pixels between title and subtitle text.

Or: ExprRef

zindex
The integer z-index indicating the layering of the title group relative to other axis, mark, and legend groups.

Default value: 0.

Or: ExprRef

Value

A modified Vega-Lite Spec