vl_config_view

Add view config (ViewConfig) to a vega-lite spec.

vl_config_view(
  spec,
  .object = NULL,
  clip = NULL,
  continuousHeight = NULL,
  continuousWidth = NULL,
  cornerRadius = NULL,
  cursor = NULL,
  discreteHeight = NULL,
  discreteWidth = NULL,
  fill = NULL,
  fillOpacity = NULL,
  opacity = NULL,
  step = NULL,
  stroke = NULL,
  strokeCap = NULL,
  strokeDash = NULL,
  strokeDashOffset = NULL,
  strokeJoin = NULL,
  strokeMiterLimit = NULL,
  strokeOpacity = NULL,
  strokeWidth = NULL
)

Arguments

spec
An input vega-lite spec

.object
Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: ViewConfig

clip
Whether the view should be clipped.

continuousHeight
The default height when the plot has a continuous y-field for x or latitude, or has arc marks.

Default value: 200

continuousWidth
The default width when the plot has a continuous field for x or longitude, or has arc marks.

Default value: 200

cornerRadius
The radius in pixels of rounded rectangles or arcs' corners.

Default value: 0

Or: ExprRef

cursor
The mouse cursor used over the view. Any valid CSS cursor type can be used.

discreteHeight
The default height when the plot has non arc marks and either a discrete y-field or no y-field. The height can be either a number indicating a fixed height or an object in the form of {step: number} defining the height per discrete step.

Default value: a step size based on config.view.step.

discreteWidth
The default width when the plot has non-arc marks and either a discrete x-field or no x-field. The width can be either a number indicating a fixed width or an object in the form of {step: number} defining the width per discrete step.

Default value: a step size based on config.view.step.

fill
The fill color.

Default value: undefined

fillOpacity
The fill opacity (value between [0,1]).

Default value: 1

Or: ExprRef

opacity
The overall opacity (value between [0,1]).

Default value: 0.7 for non-aggregate plots with point, tick, circle, or square marks or layered bar charts and 1 otherwise.

step
Default step size for x-/y- discrete fields.

stroke
The stroke color.

Default value: "#ddd"

strokeCap
The stroke cap for line ending style. One of "butt", "round", or "square".

Default value: "butt"

Or: ExprRef

strokeDash
An array of alternating stroke, space lengths for creating dashed or dotted lines.

Or: ExprRef

strokeDashOffset
The offset (in pixels) into which to begin drawing with the stroke dash array.

Or: ExprRef

strokeJoin
The stroke line join method. One of "miter", "round" or "bevel".

Default value: "miter"

Or: ExprRef

strokeMiterLimit
The miter limit at which to bevel a line join.

Or: ExprRef

strokeOpacity
The stroke opacity (value between [0,1]).

Default value: 1

Or: ExprRef

strokeWidth
The stroke width, in pixels.

Or: ExprRef

Value

A modified Vega-Lite Spec