vl_legend_size

Add legend to encoding for size in a vega-lite spec.

vl_legend_size(spec, clipHeight = NULL, columnPadding = NULL,
  columns = NULL, cornerRadius = NULL, direction = NULL,
  fillColor = NULL, format = NULL, formatType = NULL,
  gradientLength = NULL, gradientOpacity = NULL,
  gradientStrokeColor = NULL, gradientStrokeWidth = NULL,
  gradientThickness = NULL, gridAlign = NULL, labelAlign = NULL,
  labelBaseline = NULL, labelColor = NULL, labelFont = NULL,
  labelFontSize = NULL, labelFontStyle = NULL,
  labelFontWeight = NULL, labelLimit = NULL, labelOffset = NULL,
  labelOpacity = NULL, labelOverlap = NULL, labelPadding = NULL,
  labelSeparation = NULL, offset = NULL, orient = NULL,
  padding = NULL, rowPadding = NULL, strokeColor = NULL,
  symbolDash = NULL, symbolDashOffset = NULL, symbolFillColor = NULL,
  symbolOffset = NULL, symbolOpacity = NULL, symbolSize = NULL,
  symbolStrokeColor = NULL, symbolStrokeWidth = NULL,
  symbolType = NULL, tickCount = NULL, tickMinStep = NULL,
  title = NULL, titleAlign = NULL, titleAnchor = NULL,
  titleBaseline = NULL, titleColor = NULL, titleFont = NULL,
  titleFontSize = NULL, titleFontStyle = NULL,
  titleFontWeight = NULL, titleLimit = NULL, titleOpacity = NULL,
  titleOrient = NULL, titlePadding = NULL, type = NULL,
  values = NULL, zindex = NULL)

Arguments

spec
A vega-lite spec.

clipHeight
(Legend) The height in pixels to clip symbol legend entries and limit their size.

columnPadding
(Legend) The horizontal padding in pixels between symbol legend entries.

Default value: 10.

columns
(Legend) The number of columns in which to arrange symbol legend entries. A value of 0 or lower indicates a single row with one column per entry.

cornerRadius
(Legend) Corner radius for the full legend.

direction
(Legend) The direction of the legend, one of "vertical" or "horizontal".

Default value:

For top-/bottom-oriented legends, "horizontal"

For left-/right-oriented legends, "vertical"

For top/bottom-left/right-oriented legends, "horizontal" for gradient legends and "vertical" for symbol legends.

fillColor
(Legend) Background fill color for the full legend.

format
(Legend) The text formatting pattern for labels of guides (axes, legends, headers) and text marks.

If the format type is "number" (e.g., for quantitative fields), this is D3's number format pattern.

If the format type is "time" (e.g., for temporal fields), this is D3's time format pattern.

See the format documentation for more examples.

Default value: Derived from numberFormat config for number format and from timeFormat config for time format.

formatType
(Legend) The format type for labels ("number" or "time").

Default value:

"time" for temporal fields and ordinal and nomimal fields with timeUnit.

"number" for quantitative fields as well as ordinal and nomimal fields without timeUnit.

gradientLength
(Legend) The length in pixels of the primary axis of a color gradient. This value corresponds to the height of a vertical gradient or the width of a horizontal gradient.

Default value: 200.

gradientOpacity
(Legend) Opacity of the color gradient.

gradientStrokeColor
(Legend) The color of the gradient stroke, can be in hex color code or regular color name.

Default value: "lightGray".

gradientStrokeWidth
(Legend) The width of the gradient stroke, in pixels.

Default value: 0.

gradientThickness
(Legend) The thickness in pixels of the color gradient. This value corresponds to the width of a vertical gradient or the height of a horizontal gradient.

Default value: 16.

gridAlign
(Legend) The alignment to apply to symbol legends rows and columns. The supported string values are "all", "each" (the default), and none. For more information, see the grid layout documentation.

Default value: "each".

labelAlign
(Legend) The alignment of the legend label, can be left, center, or right.

labelBaseline
(Legend) The position of the baseline of legend label, can be "top", "middle", "bottom", or "alphabetic".

Default value: "middle".

labelColor
(Legend) The color of the legend label, can be in hex color code or regular color name.

labelFont
(Legend) The font of the legend label.

labelFontSize
(Legend) The font size of legend label.

Default value: 10.

labelFontStyle
(Legend) The font style of legend label.

labelFontWeight
(Legend) The font weight of legend label.

labelLimit
(Legend) Maximum allowed pixel width of axis tick labels.

Default value: 160.

labelOffset
(Legend) The offset of the legend label.

labelOpacity
(Legend) Opacity of labels.

labelOverlap
(Legend) The strategy to use for resolving overlap of labels in gradient legends. If false, no overlap reduction is attempted. If set to true (default) or "parity", a strategy of removing every other label is used. If set to "greedy", a linear scan of the labels is performed, removing any label that overlaps with the last visible label (this often works better for log-scaled axes).

Default value: true.

labelPadding
(Legend) Padding in pixels between the legend and legend labels.

labelSeparation
(Legend) The minimum separation that must be between label bounding boxes for them to be considered non-overlapping (default 0). This property is ignored if labelOverlap resolution is not enabled.

offset
(Legend) The offset in pixels by which to displace the legend from the data rectangle and axes.

Default value: 18.

orient
(Legend) The orientation of the legend, which determines how the legend is positioned within the scene. One of "left", "right", "top-left", "top-right", "bottom-left", "bottom-right", "none".

Default value: "right"

padding
(Legend) The padding between the border and content of the legend group.

Default value: 0.

rowPadding
(Legend) The vertical padding in pixels between symbol legend entries.

Default value: 2.

strokeColor
(Legend) Border stroke color for the full legend.

symbolDash
(Legend) An array of alternating stroke, space lengths for dashed symbol strokes.

symbolDashOffset
(Legend) The pixel offset at which to start drawing with the symbol stroke dash array.

symbolFillColor
(Legend) The color of the legend symbol,

symbolOffset
(Legend) Horizontal pixel offset for legend symbols.

Default value: 0.

symbolOpacity
(Legend) Opacity of the legend symbols.

symbolSize
(Legend) The size of the legend symbol, in pixels.

Default value: 100.

symbolStrokeColor
(Legend) Stroke color for legend symbols.

symbolStrokeWidth
(Legend) The width of the symbol's stroke.

Default value: 1.5.

symbolType
(Legend) Default shape type (such as "circle") for legend symbols. Can be one of ``"circle","square"`, `"cross"`, `"diamond"`, `"triangle-up"`, `"triangle-down"`, `"triangle-right"`, or `"triangle-left"`.

In addition to a set of built-in shapes, custom shapes can be defined using SVG path strings.

Default value: "circle".

tickCount
(Legend) The desired number of tick values for quantitative legends.

tickMinStep
(Legend) The minimum desired step between legend ticks, in terms of scale domain values. For example, a value of 1 indicates that ticks should not be less than 1 unit apart. If tickMinStep is specified, the tickCount value will be adjusted, if necessary, to enforce the minimum step value.

Default value: undefined

title
(Legend) A title for the field. If null, the title will be removed.

Default value: derived from the field's name and transformation function (aggregate, bin and timeUnit). If the field has an aggregate function, the function is displayed as part of the title (e.g., "Sum of Profit"). If the field is binned or has a time unit applied, the applied function is shown in parentheses (e.g., "Profit (binned)", "Transaction Date (year-month)"). Otherwise, the title is simply the field name.

Notes:

You can customize the default field title format by providing the fieldTitle property in the config or fieldTitle function via the compile function's options.

If both field definition's title and axis, header, or legend title are defined, axis/header/legend title will be used.

titleAlign
(Legend) Horizontal text alignment for legend titles.

Default value: "left".

titleAnchor
(Legend) Text anchor position for placing legend titles.

titleBaseline
(Legend) Vertical text baseline for legend titles.

Default value: "top".

titleColor
(Legend) The color of the legend title, can be in hex color code or regular color name.

titleFont
(Legend) The font of the legend title.

titleFontSize
(Legend) The font size of the legend title.

titleFontStyle
(Legend) The font style of the legend title.

titleFontWeight
(Legend) The font weight of the legend title. This can be either a string (e.g "bold", "normal") or a number (100, 200, 300, ..., 900 where "normal" = 400 and "bold" = 700).

titleLimit
(Legend) Maximum allowed pixel width of axis titles.

Default value: 180.

titleOpacity
(Legend) Opacity of the legend title.

titleOrient
(Legend) Orientation of the legend title.

titlePadding
(Legend) The padding, in pixels, between title and legend.

Default value: 5.

type
(Legend) The type of the legend. Use "symbol" to create a discrete legend and "gradient" for a continuous color gradient.

Default value: "gradient" for non-binned quantitative fields and temporal fields; "symbol" otherwise.

values
(Legend) Explicitly set the visible legend values.

zindex
(Legend) A non-positive integer indicating z-index of the legend. If zindex is 0, legend should be drawn behind all chart elements. To put them in front, use zindex = 1.

Value

A modified spec

See also