vl_make_LegendConfig

Create spec for obj

vl_make_LegendConfig(clipHeight = NULL, columnPadding = NULL,
  columns = NULL, cornerRadius = NULL, fillColor = NULL,
  gradientDirection = NULL, gradientHorizontalMaxLength = NULL,
  gradientHorizontalMinLength = NULL, gradientLabelLimit = NULL,
  gradientLabelOffset = NULL, gradientLength = NULL,
  gradientOpacity = NULL, gradientStrokeColor = NULL,
  gradientStrokeWidth = NULL, gradientThickness = NULL,
  gradientVerticalMaxLength = NULL, gradientVerticalMinLength = NULL,
  gridAlign = NULL, labelAlign = NULL, labelBaseline = NULL,
  labelColor = NULL, labelFont = NULL, labelFontSize = NULL,
  labelFontStyle = NULL, labelFontWeight = NULL, labelLimit = NULL,
  labelOffset = NULL, labelOpacity = NULL, labelOverlap = NULL,
  labelPadding = NULL, labelSeparation = NULL, layout = NULL,
  legendX = NULL, legendY = NULL, offset = NULL, orient = NULL,
  padding = NULL, rowPadding = NULL, shortTimeLabels = NULL,
  strokeColor = NULL, strokeDash = NULL, strokeWidth = NULL,
  symbolBaseFillColor = NULL, symbolBaseStrokeColor = NULL,
  symbolDash = NULL, symbolDashOffset = NULL, symbolDirection = NULL,
  symbolFillColor = NULL, symbolOffset = NULL, symbolOpacity = NULL,
  symbolSize = NULL, symbolStrokeColor = NULL,
  symbolStrokeWidth = NULL, symbolType = NULL, title = NULL,
  titleAlign = NULL, titleAnchor = NULL, titleBaseline = NULL,
  titleColor = NULL, titleFont = NULL, titleFontSize = NULL,
  titleFontStyle = NULL, titleFontWeight = NULL, titleLimit = NULL,
  titleOpacity = NULL, titleOrient = NULL, titlePadding = NULL)

Arguments

clipHeight
(LegendConfig) The height in pixels to clip symbol legend entries and limit their size.

columnPadding
(LegendConfig) The horizontal padding in pixels between symbol legend entries.

Default value: 10.

columns
(LegendConfig) The number of columns in which to arrange symbol legend entries. A value of 0 or lower indicates a single row with one column per entry.

cornerRadius
(LegendConfig) Corner radius for the full legend.

fillColor
(LegendConfig) Background fill color for the full legend.

gradientDirection
(LegendConfig) The default direction ("horizontal" or "vertical") for gradient legends.

Default value: "vertical".

gradientHorizontalMaxLength
(LegendConfig) Max legend length for a horizontal gradient when config.legend.gradientLength is undefined.

Default value: 200

gradientHorizontalMinLength
(LegendConfig) Min legend length for a horizontal gradient when config.legend.gradientLength is undefined.

Default value: 100

gradientLabelLimit
(LegendConfig) The maximum allowed length in pixels of color ramp gradient labels.

gradientLabelOffset
(LegendConfig) Vertical offset in pixels for color ramp gradient labels.

Default value: 2.

gradientLength
(LegendConfig) The length in pixels of the primary axis of a color gradient. This value corresponds to the height of a vertical gradient or the width of a horizontal gradient.

Default value: 200.

gradientOpacity
(LegendConfig) Opacity of the color gradient.

gradientStrokeColor
(LegendConfig) The color of the gradient stroke, can be in hex color code or regular color name.

Default value: "lightGray".

gradientStrokeWidth
(LegendConfig) The width of the gradient stroke, in pixels.

Default value: 0.

gradientThickness
(LegendConfig) The thickness in pixels of the color gradient. This value corresponds to the width of a vertical gradient or the height of a horizontal gradient.

Default value: 16.

gradientVerticalMaxLength
(LegendConfig) Max legend length for a vertical gradient when config.legend.gradientLength is undefined.

Default value: 200

gradientVerticalMinLength
(LegendConfig) Min legend length for a vertical gradient when config.legend.gradientLength is undefined.

Default value: 100

gridAlign
(LegendConfig) The alignment to apply to symbol legends rows and columns. The supported string values are "all", "each" (the default), and none. For more information, see the grid layout documentation.

Default value: "each".

labelAlign
(LegendConfig) The alignment of the legend label, can be left, center, or right.

labelBaseline
(LegendConfig) The position of the baseline of legend label, can be "top", "middle", "bottom", or "alphabetic".

Default value: "middle".

labelColor
(LegendConfig) The color of the legend label, can be in hex color code or regular color name.

labelFont
(LegendConfig) The font of the legend label.

labelFontSize
(LegendConfig) The font size of legend label.

Default value: 10.

labelFontStyle
(LegendConfig) The font style of legend label.

labelFontWeight
(LegendConfig) The font weight of legend label.

labelLimit
(LegendConfig) Maximum allowed pixel width of axis tick labels.

Default value: 160.

labelOffset
(LegendConfig) The offset of the legend label.

labelOpacity
(LegendConfig) Opacity of labels.

labelOverlap
(LegendConfig) The strategy to use for resolving overlap of labels in gradient legends. If false, no overlap reduction is attempted. If set to true or "parity", a strategy of removing every other label is used. If set to "greedy", a linear scan of the labels is performed, removing any label that overlaps with the last visible label (this often works better for log-scaled axes).

Default value: "greedy" for log scales otherwise true`. *

labelPadding
(LegendConfig) Padding in pixels between the legend and legend labels.

labelSeparation
(LegendConfig) The minimum separation that must be between label bounding boxes for them to be considered non-overlapping (default 0). This property is ignored if labelOverlap resolution is not enabled.

layout
(LegendConfig) Legend orient group layout parameters.

legendX
(LegendConfig) Custom x-position for legend with orient "none".

legendY
(LegendConfig) Custom y-position for legend with orient "none".

offset
(LegendConfig) The offset in pixels by which to displace the legend from the data rectangle and axes.

Default value: 18.

orient
(LegendConfig) The orientation of the legend, which determines how the legend is positioned within the scene. One of "left", "right", "top-left", "top-right", "bottom-left", "bottom-right", "none".

Default value: "right"

padding
(LegendConfig) The padding between the border and content of the legend group.

Default value: 0.

rowPadding
(LegendConfig) The vertical padding in pixels between symbol legend entries.

Default value: 2.

shortTimeLabels
(LegendConfig) Whether month names and weekday names should be abbreviated.

Default value: false

strokeColor
(LegendConfig) Border stroke color for the full legend.

strokeDash
(LegendConfig) Border stroke dash pattern for the full legend.

strokeWidth
(LegendConfig) Border stroke width for the full legend.

symbolBaseFillColor
(LegendConfig) Default fill color for legend symbols. Only applied if there is no "fill" scale color encoding for the legend.

Default value: "transparent".

symbolBaseStrokeColor
(LegendConfig) Default stroke color for legend symbols. Only applied if there is no "fill" scale color encoding for the legend.

Default value: "gray".

symbolDash
(LegendConfig) An array of alternating [stroke, space] lengths for dashed symbol strokes.

symbolDashOffset
(LegendConfig) The pixel offset at which to start drawing with the symbol stroke dash array.

symbolDirection
(LegendConfig) The default direction ("horizontal" or "vertical") for symbol legends.

Default value: "vertical".

symbolFillColor
(LegendConfig) The color of the legend symbol,

symbolOffset
(LegendConfig) Horizontal pixel offset for legend symbols.

Default value: 0.

symbolOpacity
(LegendConfig) Opacity of the legend symbols.

symbolSize
(LegendConfig) The size of the legend symbol, in pixels.

Default value: 100.

symbolStrokeColor
(LegendConfig) Stroke color for legend symbols.

symbolStrokeWidth
(LegendConfig) The width of the symbol's stroke.

Default value: 1.5.

symbolType
(LegendConfig) Default shape type (such as "circle") for legend symbols. Can be one of ``"circle", "square"`, `"cross"`, `"diamond"`, `"triangle-up"`, `"triangle-down"`, `"triangle-right"`, or `"triangle-left"`.

In addition to a set of built-in shapes, custom shapes can be defined using SVG path strings.

Default value: "circle".

title
(LegendConfig) Set to null to disable title for the axis, legend, or header.

titleAlign
(LegendConfig) Horizontal text alignment for legend titles.

Default value: "left".

titleAnchor
(LegendConfig) Text anchor position for placing legend titles.

titleBaseline
(LegendConfig) Vertical text baseline for legend titles.

Default value: "top".

titleColor
(LegendConfig) The color of the legend title, can be in hex color code or regular color name.

titleFont
(LegendConfig) The font of the legend title.

titleFontSize
(LegendConfig) The font size of the legend title.

titleFontStyle
(LegendConfig) The font style of the legend title.

titleFontWeight
(LegendConfig) The font weight of the legend title. This can be either a string (e.g "bold", "normal") or a number (100, 200, 300, ..., 900 where "normal" = 400 and "bold" = 700).

titleLimit
(LegendConfig) Maximum allowed pixel width of axis titles.

Default value: 180.

titleOpacity
(LegendConfig) Opacity of the legend title.

titleOrient
(LegendConfig) Orientation of the legend title.

titlePadding
(LegendConfig) The padding, in pixels, between title and legend.

Default value: 5.

Value

A component of a Vega-Lite spec, corresponding to a LegendConfig.