vl_make_ScaleConfig

Create spec for ScaleConfig.

vl_make_ScaleConfig(bandPaddingInner = NULL, bandPaddingOuter = NULL,
  barBandPaddingInner = NULL, barBandPaddingOuter = NULL,
  clamp = NULL, continuousPadding = NULL, maxBandSize = NULL,
  maxFontSize = NULL, maxOpacity = NULL, maxSize = NULL,
  maxStrokeWidth = NULL, minBandSize = NULL, minFontSize = NULL,
  minOpacity = NULL, minSize = NULL, minStrokeWidth = NULL,
  pointPadding = NULL, quantileCount = NULL, quantizeCount = NULL,
  rangeStep = NULL, rectBandPaddingInner = NULL,
  rectBandPaddingOuter = NULL, round = NULL, textXRangeStep = NULL,
  useUnaggregatedDomain = NULL)

Arguments

bandPaddingInner
(ScaleConfig) Default inner padding for x and y band-ordinal scales.

Default value:

barBandPaddingInner for bar marks (0.1 by default)

rectBandPaddingInner for rect and other marks (0 by default)

bandPaddingOuter
(ScaleConfig) Default outer padding for x and y band-ordinal scales.

If not specified, by default, band scale's paddingOuter is paddingInner/2.

barBandPaddingInner
(ScaleConfig) Default inner padding for x and y band-ordinal scales of "bar" marks.

Default value: 0.1

barBandPaddingOuter
(ScaleConfig) Default outer padding for x and y band-ordinal scales of "bar" marks. If not specified, by default, band scale's paddingOuter is paddingInner/2.

clamp
(ScaleConfig) If true, values that exceed the data domain are clamped to either the minimum or maximum range value

continuousPadding
(ScaleConfig) Default padding for continuous scales.

Default: 5 for continuous x-scale of a vertical bar and continuous y-scale of a horizontal bar.; 0 otherwise.

maxBandSize
(ScaleConfig) The default max value for mapping quantitative fields to bar's size/bandSize.

If undefined (default), we will use the scale's rangeStep - 1.

maxFontSize
(ScaleConfig) The default max value for mapping quantitative fields to text's size/fontSize.

Default value: 40

maxOpacity
(ScaleConfig) Default max opacity for mapping a field to opacity.

Default value: 0.8

maxSize
(ScaleConfig) Default max value for point size scale.

maxStrokeWidth
(ScaleConfig) Default max strokeWidth for the scale of strokeWidth for rule and line marks and of size for trail marks.

Default value: 4

minBandSize
(ScaleConfig) The default min value for mapping quantitative fields to bar and tick's size/bandSize scale with zero=false.

Default value: 2

minFontSize
(ScaleConfig) The default min value for mapping quantitative fields to tick's size/fontSize scale with zero=false

Default value: 8

minOpacity
(ScaleConfig) Default minimum opacity for mapping a field to opacity.

Default value: 0.3

minSize
(ScaleConfig) Default minimum value for point size scale with zero=false.

Default value: 9

minStrokeWidth
(ScaleConfig) Default minimum strokeWidth for the scale of strokeWidth for rule and line marks and of size for trail marks with zero=false.

Default value: 1

pointPadding
(ScaleConfig) Default outer padding for x and y point-ordinal scales.

Default value: 0.5

quantileCount
(ScaleConfig) Default range cardinality for quantile scale.

Default value: 4

quantizeCount
(ScaleConfig) Default range cardinality for quantize scale.

Default value: 4

rangeStep
(ScaleConfig) Default range step for band and point scales of (1) the y channel and (2) the x channel when the mark is not text.

Default value: 20

rectBandPaddingInner
(ScaleConfig) Default inner padding for x and y band-ordinal scales of "rect" marks.

Default value: 0

rectBandPaddingOuter
(ScaleConfig) Default outer padding for x and y band-ordinal scales of "rect" marks. If not specified, by default, band scale's paddingOuter is paddingInner/2.

round
(ScaleConfig) If true, rounds numeric output values to integers. This can be helpful for snapping to the pixel grid. (Only available for x, y, and size scales.)

textXRangeStep
(ScaleConfig) Default range step for x band and point scales of text marks.

Default value: 90

useUnaggregatedDomain
(ScaleConfig) Use the source data range before aggregation as scale domain instead of aggregated data for aggregate axis.

This is equivalent to setting domain to "unaggregate" for aggregated quantitative fields by default.

This property only works with aggregate functions that produce values within the raw data domain ("mean", "average", "median", "q1", "q3", "min", "max"). For other aggregations that produce values outside of the raw data domain (e.g. "count", "sum"), this property is ignored.

Default value: false

Value

A modified spec