vl_mark_errorband

vl_mark_errorband

vl_mark_errorband(spec, band = NULL, borders = NULL, clip = NULL,
  color = NULL, extent = NULL, interpolate = NULL, opacity = NULL,
  orient = NULL, tension = NULL)

Arguments

spec
An input vega-lite spec

band
(ErrorBandDef)

borders
(ErrorBandDef)

clip
(ErrorBandDef) Whether a composite mark be clipped to the enclosing group’s width and height.

color
(ErrorBandDef) Default color. Note that fill and stroke have higher precedence than color and will override color.

Default value: <span style="color: #4682b4;">■</span> "#4682b4"

Note: This property cannot be used in a style config.

extent
(ErrorBandDef) The extent of the band. Available options include:

"ci": Extend the band to the confidence interval of the mean.

"stderr": The size of band are set to the value of standard error, extending from the mean.

"stdev": The size of band are set to the value of standard deviation, extending from the mean.

"iqr": Extend the band to the q1 and q3.

Default value: "stderr".

interpolate
(ErrorBandDef) The line interpolation method for the error band. One of the following:

"linear": piecewise linear segments, as in a polyline.

"linear-closed": close the linear segments to form a polygon.

"step": a piecewise constant function (a step function) consisting of alternating horizontal and vertical lines. The y-value changes at the midpoint of each pair of adjacent x-values.

"step-before": a piecewise constant function (a step function) consisting of alternating horizontal and vertical lines. The y-value changes before the x-value.

"step-after": a piecewise constant function (a step function) consisting of alternating horizontal and vertical lines. The y-value changes after the x-value.

"basis": a B-spline, with control point duplication on the ends.

"basis-open": an open B-spline; may not intersect the start or end.

"basis-closed": a closed B-spline, as in a loop.

"cardinal": a Cardinal spline, with control point duplication on the ends.

"cardinal-open": an open Cardinal spline; may not intersect the start or end, but will intersect other control points.

"cardinal-closed": a closed Cardinal spline, as in a loop.

"bundle": equivalent to basis, except the tension parameter is used to straighten the spline.

"monotone": cubic interpolation that preserves monotonicity in y.

opacity
(ErrorBandDef) The opacity (value between [0,1]) of the mark.

orient
(ErrorBandDef) Orientation of the error band. This is normally automatically determined, but can be specified when the orientation is ambiguous and cannot be automatically determined.

tension
(ErrorBandDef) The tension parameter for the interpolation type of the error band.

Value

A modified Vega-Lite Spec