vl_mark_rect

vl_mark_rect

Usage

vl_mark_rect(
  spec,
  align = NULL,
  angle = NULL,
  aria = NULL,
  ariaRole = NULL,
  ariaRoleDescription = NULL,
  aspect = NULL,
  bandSize = NULL,
  baseline = NULL,
  binSpacing = NULL,
  blend = NULL,
  clip = NULL,
  color = NULL,
  continuousBandSize = NULL,
  cornerRadius = NULL,
  cornerRadiusBottomLeft = NULL,
  cornerRadiusBottomRight = NULL,
  cornerRadiusEnd = NULL,
  cornerRadiusTopLeft = NULL,
  cornerRadiusTopRight = NULL,
  cursor = NULL,
  description = NULL,
  dir = NULL,
  discreteBandSize = NULL,
  dx = NULL,
  dy = NULL,
  ellipsis = NULL,
  fill = NULL,
  fillOpacity = NULL,
  filled = NULL,
  font = NULL,
  fontSize = NULL,
  fontStyle = NULL,
  fontWeight = NULL,
  height = NULL,
  href = NULL,
  innerRadius = NULL,
  interpolate = NULL,
  invalid = NULL,
  limit = NULL,
  line = NULL,
  lineBreak = NULL,
  lineHeight = NULL,
  opacity = NULL,
  order = NULL,
  orient = NULL,
  outerRadius = NULL,
  padAngle = NULL,
  point = NULL,
  radius = NULL,
  radius2 = NULL,
  radius2Offset = NULL,
  radiusOffset = NULL,
  shape = NULL,
  size = NULL,
  smooth = NULL,
  stroke = NULL,
  strokeCap = NULL,
  strokeDash = NULL,
  strokeDashOffset = NULL,
  strokeJoin = NULL,
  strokeMiterLimit = NULL,
  strokeOffset = NULL,
  strokeOpacity = NULL,
  strokeWidth = NULL,
  style = NULL,
  tension = NULL,
  text = NULL,
  theta = NULL,
  theta2 = NULL,
  theta2Offset = NULL,
  thetaOffset = NULL,
  thickness = NULL,
  timeUnitBandPosition = NULL,
  timeUnitBandSize = NULL,
  tooltip = NULL,
  url = NULL,
  width = NULL,
  x = NULL,
  x2 = NULL,
  x2Offset = NULL,
  xOffset = NULL,
  y = NULL,
  y2 = NULL,
  y2Offset = NULL,
  yOffset = NULL,
  .object = NULL
)

Arguments

spec

An input vega-lite spec

align

The horizontal alignment of the text or ranged marks (area, bar, image, rect, rule). One of "left", "right", "center".

Note: Expression reference is not supported for range marks.

angle

The rotation angle of the text, in degrees.

Or: ExprRef

aria

A boolean flag indicating if ARIA attributes should be included (SVG output only). If false, the "aria-hidden" attribute will be set on the output SVG element, removing the mark item from the ARIA accessibility tree.

Or: ExprRef

ariaRole

Sets the type of user interface element of the mark item for ARIA accessibility (SVG output only). If specified, this property determines the "role" attribute. Warning: this property is experimental and may be changed in the future.

Or: ExprRef

ariaRoleDescription

A human-readable, author-localized description for the role of the mark item for ARIA accessibility (SVG output only). If specified, this property determines the "aria-roledescription" attribute. Warning: this property is experimental and may be changed in the future.

Or: ExprRef

aspect

Whether to keep aspect ratio of image marks.

Or: ExprRef

bandSize

The width of the ticks.

Default value: 3/4 of step (width step for horizontal ticks and height step for vertical ticks).

baseline

For text marks, the vertical text baseline. One of "alphabetic" (default), "top", "middle", "bottom", "line-top", "line-bottom", or an expression reference that provides one of the valid values. The "line-top" and "line-bottom" values operate similarly to "top" and "bottom", but are calculated relative to the lineHeight rather than fontSize alone.

For range marks, the vertical alignment of the marks. One of "top", "middle", "bottom".

Note: Expression reference is not supported for range marks.

binSpacing

Offset between bars for binned field. The ideal value for this is either 0 (preferred by statisticians) or 1 (Vega-Lite default, D3 example style).

Default value: 1

blend

The color blend mode for drawing an item on its current background. Any valid CSS mix-blend-mode value can be used.

__Default value: "source-over"

Or: ExprRef

clip

Whether a mark be clipped to the enclosing group’s width and height.

color

Default color.

Default value: <span style="color: #4682b4;">■</span> "#4682b4"

Note:

This property cannot be used in a style config.

The fill and stroke properties have higher precedence than color and will override color.

continuousBandSize

The default size of the bars on continuous scales.

Default value: 5

cornerRadius

The radius in pixels of rounded rectangles or arcs' corners.

Default value: 0

Or: ExprRef

cornerRadiusBottomLeft

The radius in pixels of rounded rectangles' bottom left corner.

Default value: 0

Or: ExprRef

cornerRadiusBottomRight

The radius in pixels of rounded rectangles' bottom right corner.

Default value: 0

Or: ExprRef

cornerRadiusEnd

For vertical bars, top-left and top-right corner radius.

For horizontal bars, top-right and bottom-right corner radius.

cornerRadiusTopLeft

The radius in pixels of rounded rectangles' top right corner.

Default value: 0

Or: ExprRef

cornerRadiusTopRight

The radius in pixels of rounded rectangles' top left corner.

Default value: 0

Or: ExprRef

cursor

The mouse cursor used over the mark. Any valid CSS cursor type can be used.

Or: ExprRef

description

A text description of the mark item for ARIA accessibility (SVG output only). If specified, this property determines the "aria-label" attribute.

Or: ExprRef

dir

The direction of the text. One of "ltr" (left-to-right) or "rtl" (right-to-left). This property determines on which side is truncated in response to the limit parameter.

Default value: "ltr"

Or: ExprRef

discreteBandSize

The default size of the bars with discrete dimensions. If unspecified, the default size is step-2, which provides 2 pixel offset between bars.

dx

The horizontal offset, in pixels, between the text label and its anchor point. The offset is applied after rotation by the angle property.

Or: ExprRef

dy

The vertical offset, in pixels, between the text label and its anchor point. The offset is applied after rotation by the angle property.

Or: ExprRef

ellipsis

The ellipsis string for text truncated in response to the limit parameter.

Default value: "…"

Or: ExprRef

fill

Default fill color. This property has higher precedence than config.color. Set to null to remove fill.

Default value: (None)

fillOpacity

The fill opacity (value between [0,1]).

Default value: 1

Or: ExprRef

filled

Whether the mark's color should be used as fill color instead of stroke color.

Default value: false for all point, line, and rule marks as well as geoshape marks for graticule data sources; otherwise, true.

Note: This property cannot be used in a style config.

font

The typeface to set the text in (e.g., "Helvetica Neue").

Or: ExprRef

fontSize

The font size, in pixels.

Default value: 11

Or: ExprRef

fontStyle

The font style (e.g., "italic").

Or: ExprRef

fontWeight

The font weight. This can be either a string (e.g "bold", "normal") or a number (100, 200, 300, ..., 900 where "normal" = 400 and "bold" = 700).

Or: ExprRef

height

Height of the marks. One of:

A number representing a fixed pixel height.

A relative band size definition. For example, {band: 0.5} represents half of the band

href

A URL to load upon mouse click. If defined, the mark acts as a hyperlink.

Or: ExprRef

innerRadius

The inner radius in pixels of arc marks. innerRadius is an alias for radius2.

Default value: 0

interpolate

The line interpolation method to use for line and area marks. One of the following:

"linear": piecewise linear segments, as in a polyline.

"linear-closed": close the linear segments to form a polygon.

"step": alternate between horizontal and vertical segments, as in a step function.

"step-before": alternate between vertical and horizontal segments, as in a step function.

"step-after": alternate between horizontal and vertical segments, as in a step function.

"basis": a B-spline, with control point duplication on the ends.

"basis-open": an open B-spline; may not intersect the start or end.

"basis-closed": a closed B-spline, as in a loop.

"cardinal": a Cardinal spline, with control point duplication on the ends.

"cardinal-open": an open Cardinal spline; may not intersect the start or end, but will intersect other control points.

"cardinal-closed": a closed Cardinal spline, as in a loop.

"bundle": equivalent to basis, except the tension parameter is used to straighten the spline.

"monotone": cubic interpolation that preserves monotonicity in y.

Or: ExprRef

invalid

Defines how Vega-Lite should handle marks for invalid values (null and NaN).

If set to "filter" (default), all data items with null values will be skipped (for line, trail, and area marks) or filtered (for other marks).

If null, all data items are included. In this case, invalid values will be interpreted as zeroes.

limit

The maximum length of the text mark in pixels. The text value will be automatically truncated if the rendered size exceeds the limit.

Default value: 0 -- indicating no limit

Or: ExprRef

line

A flag for overlaying line on top of area marks, or an object defining the properties of the overlayed lines.

If this value is an empty object ({}) or true, lines with default properties will be used.

If this value is false, no lines would be automatically added to area marks.

Default value: false.

lineBreak

A delimiter, such as a newline character, upon which to break text strings into multiple lines. This property is ignored if the text is array-valued.

Or: ExprRef

lineHeight

The line height in pixels (the spacing between subsequent lines of text) for multi-line text marks.

Or: ExprRef

opacity

The overall opacity (value between [0,1]).

Default value: 0.7 for non-aggregate plots with point, tick, circle, or square marks or layered bar charts and 1 otherwise.

order

For line and trail marks, this order property can be set to null or false to make the lines use the original order in the data sources.

orient

The orientation of a non-stacked bar, tick, area, and line charts. The value is either horizontal (default) or vertical.

For bar, rule and tick, this determines whether the size of the bar and tick should be applied to x or y dimension.

For area, this property determines the orient property of the Vega output.

For line and trail marks, this property determines the sort order of the points in the line if config.sortLineBy is not specified. For stacked charts, this is always determined by the orientation of the stack; therefore explicitly specified value will be ignored.

outerRadius

The outer radius in pixels of arc marks. outerRadius is an alias for radius.

Default value: 0

padAngle

The angular padding applied to sides of the arc, in radians.

Or: ExprRef

point

A flag for overlaying points on top of line or area marks, or an object defining the properties of the overlayed points.

If this property is "transparent", transparent points will be used (for enhancing tooltips and selections).

If this property is an empty object ({}) or true, filled points with default properties will be used.

If this property is false, no points would be automatically added to line or area marks.

Default value: false.

radius

For arc mark, the primary (outer) radius in pixels.

For text marks, polar coordinate radial offset, in pixels, of the text from the origin determined by the x and y properties.

Default value: min(plot_width, plot_height)/2

radius2

The secondary (inner) radius in pixels of arc marks.

Default value: 0

radius2Offset

Offset for radius2.

radiusOffset

Offset for radius.

shape

Shape of the point marks. Supported values include:

plotting shapes: "circle", "square", "cross", "diamond", "triangle-up", "triangle-down", "triangle-right", or "triangle-left".

the line symbol "stroke"

centered directional shapes "arrow", "wedge", or "triangle"

a custom SVG path string (For correct sizing, custom shape paths should be defined within a square bounding box with coordinates ranging from -1 to 1 along both the x and y dimensions.)

Default value: "circle"

Or: ExprRef

size

Default size for marks.

For point/circle/square, this represents the pixel area of the marks. Note that this value sets the area of the symbol; the side lengths will increase with the square root of this value.

For bar, this represents the band size of the bar, in pixels.

For text, this represents the font size, in pixels.

Default value:

30 for point, circle, square marks; width/height's step

2 for bar marks with discrete dimensions;

5 for bar marks with continuous dimensions;

11 for text marks.

smooth

A boolean flag (default true) indicating if the image should be smoothed when resized. If false, individual pixels should be scaled directly rather than interpolated with smoothing. For SVG rendering, this option may not work in some browsers due to lack of standardization.

Or: ExprRef

stroke

Default stroke color. This property has higher precedence than config.color. Set to null to remove stroke.

Default value: (None)

strokeCap

The stroke cap for line ending style. One of "butt", "round", or "square".

Default value: "butt"

Or: ExprRef

strokeDash

An array of alternating stroke, space lengths for creating dashed or dotted lines.

Or: ExprRef

strokeDashOffset

The offset (in pixels) into which to begin drawing with the stroke dash array.

Or: ExprRef

strokeJoin

The stroke line join method. One of "miter", "round" or "bevel".

Default value: "miter"

Or: ExprRef

strokeMiterLimit

The miter limit at which to bevel a line join.

Or: ExprRef

strokeOffset

The offset in pixels at which to draw the group stroke and fill. If unspecified, the default behavior is to dynamically offset stroked groups such that 1 pixel stroke widths align with the pixel grid.

Or: ExprRef

strokeOpacity

The stroke opacity (value between [0,1]).

Default value: 1

Or: ExprRef

strokeWidth

The stroke width, in pixels.

Or: ExprRef

style

A string or array of strings indicating the name of custom styles to apply to the mark. A style is a named collection of mark property defaults defined within the style configuration. If style is an array, later styles will override earlier styles. Any mark properties explicitly defined within the encoding will override a style default.

Default value: The mark's name. For example, a bar mark will have style "bar" by default. Note: Any specified style will augment the default style. For example, a bar mark with "style": "foo" will receive from config.style.bar and config.style.foo (the specified style "foo" has higher precedence).

tension

Depending on the interpolation type, sets the tension parameter (for line and area marks).

Or: ExprRef

text

Placeholder text if the text channel is not specified

Or: ExprRef

theta

For arc marks, the arc length in radians if theta2 is not specified, otherwise the start arc angle. (A value of 0 indicates up or “north”, increasing values proceed clockwise.)

For text marks, polar coordinate angle in radians.

theta2

The end angle of arc marks in radians. A value of 0 indicates up or “north”, increasing values proceed clockwise.

theta2Offset

Offset for theta2.

thetaOffset

Offset for theta.

thickness

Thickness of the tick mark.

Default value: 1

timeUnitBandPosition

Default relative band position for a time unit. If set to 0, the marks will be positioned at the beginning of the time unit band step. If set to 0.5, the marks will be positioned in the middle of the time unit band step.

timeUnitBandSize

Default relative band size for a time unit. If set to 1, the bandwidth of the marks will be equal to the time unit band step. If set to 0.5, bandwidth of the marks will be half of the time unit band step.

tooltip

The tooltip text string to show upon mouse hover or an object defining which fields should the tooltip be derived from.

If tooltip is true or {"content": "encoding"}, then all fields from encoding will be used.

If tooltip is {"content": "data"}, then all fields that appear in the highlighted data point will be used.

If set to null or false, then no tooltip will be used.

See the tooltip documentation for a detailed discussion about tooltip in Vega-Lite.

Default value: null

url

The URL of the image file for image marks.

Or: ExprRef

width

Width of the marks. One of:

A number representing a fixed pixel width.

A relative band size definition. For example, {band: 0.5} represents half of the band.

x

X coordinates of the marks, or width of horizontal "bar" and "area" without specified x2 or width.

The value of this channel can be a number or a string "width" for the width of the plot.

x2

X2 coordinates for ranged "area", "bar", "rect", and "rule".

The value of this channel can be a number or a string "width" for the width of the plot.

x2Offset

Offset for x2-position.

xOffset

Offset for x-position.

y

Y coordinates of the marks, or height of vertical "bar" and "area" without specified y2 or height.

The value of this channel can be a number or a string "height" for the height of the plot.

y2

Y2 coordinates for ranged "area", "bar", "rect", and "rule".

The value of this channel can be a number or a string "height" for the height of the plot.

y2Offset

Offset for y2-position.

yOffset

Offset for y-position.

.object

Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than spec. Objects can be of type: MarkDef

Value

A modified Vega-Lite Spec