vl_mark_rule

vl_mark_rule

vl_mark_rule(spec, align = NULL, angle = NULL, baseline = NULL,
  binSpacing = NULL, clip = NULL, color = NULL,
  cornerRadius = NULL, cursor = NULL, dir = NULL, dx = NULL,
  dy = NULL, ellipsis = NULL, fill = NULL, fillOpacity = NULL,
  filled = NULL, font = NULL, fontSize = NULL, fontStyle = NULL,
  fontWeight = NULL, href = NULL, interpolate = NULL, limit = NULL,
  line = NULL, opacity = NULL, order = NULL, orient = NULL,
  point = NULL, radius = NULL, shape = NULL, size = NULL,
  stroke = NULL, strokeCap = NULL, strokeDash = NULL,
  strokeDashOffset = NULL, strokeJoin = NULL,
  strokeMiterLimit = NULL, strokeOpacity = NULL, strokeWidth = NULL,
  style = NULL, tension = NULL, text = NULL, theta = NULL,
  thickness = NULL, tooltip = NULL, type = NULL, x = NULL,
  x2 = NULL, x2Offset = NULL, xOffset = NULL, y = NULL,
  y2 = NULL, y2Offset = NULL, yOffset = NULL)

Arguments

spec
An input vega-lite spec

align
(MarkDef) The horizontal alignment of the text. One of "left", "right", "center".

angle
(MarkDef) The rotation angle of the text, in degrees.

baseline
(MarkDef) The vertical alignment of the text. One of "top", "middle", "bottom".

Default value: "middle"

binSpacing
(MarkDef) Offset between bars for binned field. Ideal value for this is either 0 (Preferred by statisticians) or 1 (Vega-Lite Default, D3 example style).

Default value: 1

clip
(MarkDef) Whether a mark be clipped to the enclosing group’s width and height.

color
(MarkDef) Default color. Note that fill and stroke have higher precedence than color and will override color.

Default value: <span style="color: #4682b4;">■</span> "#4682b4"

Note: This property cannot be used in a style config.

cornerRadius
(MarkDef) The radius in pixels of rounded rectangle corners.

Default value: 0

cursor
(MarkDef) The mouse cursor used over the mark. Any valid CSS cursor type can be used.

dir
(MarkDef) The direction of the text. One of "ltr" (left-to-right) or "rtl" (right-to-left). This property determines on which side is truncated in response to the limit parameter.

Default value: "ltr"

dx
(MarkDef) The horizontal offset, in pixels, between the text label and its anchor point. The offset is applied after rotation by the angle property.

dy
(MarkDef) The vertical offset, in pixels, between the text label and its anchor point. The offset is applied after rotation by the angle property.

ellipsis
(MarkDef) The ellipsis string for text truncated in response to the limit parameter.

Default value: "…"

fill
(MarkDef) Default Fill Color. This has higher precedence than config.color

Default value: (None)

fillOpacity
(MarkDef) The fill opacity (value between [0,1]).

Default value: 1

filled
(MarkDef) Whether the mark's color should be used as fill color instead of stroke color.

Default value: false for point, line and rule; otherwise, true.

Note: This property cannot be used in a style config.

font
(MarkDef) The typeface to set the text in (e.g., "Helvetica Neue").

fontSize
(MarkDef) The font size, in pixels.

fontStyle
(MarkDef) The font style (e.g., "italic").

fontWeight
(MarkDef) The font weight. This can be either a string (e.g "bold", "normal") or a number (100, 200, 300, ..., 900 where "normal" = 400 and "bold" = 700).

href
(MarkDef) A URL to load upon mouse click. If defined, the mark acts as a hyperlink.

interpolate
(MarkDef) The line interpolation method to use for line and area marks. One of the following:

"linear": piecewise linear segments, as in a polyline.

"linear-closed": close the linear segments to form a polygon.

"step": alternate between horizontal and vertical segments, as in a step function.

"step-before": alternate between vertical and horizontal segments, as in a step function.

"step-after": alternate between horizontal and vertical segments, as in a step function.

"basis": a B-spline, with control point duplication on the ends.

"basis-open": an open B-spline; may not intersect the start or end.

"basis-closed": a closed B-spline, as in a loop.

"cardinal": a Cardinal spline, with control point duplication on the ends.

"cardinal-open": an open Cardinal spline; may not intersect the start or end, but will intersect other control points.

"cardinal-closed": a closed Cardinal spline, as in a loop.

"bundle": equivalent to basis, except the tension parameter is used to straighten the spline.

"monotone": cubic interpolation that preserves monotonicity in y.

limit
(MarkDef) The maximum length of the text mark in pixels. The text value will be automatically truncated if the rendered size exceeds the limit.

Default value: 0, indicating no limit

line
(MarkDef) A flag for overlaying line on top of area marks, or an object defining the properties of the overlayed lines.

If this value is an empty object ({}) or true, lines with default properties will be used.

If this value is false, no lines would be automatically added to area marks.

Default value: false.

opacity
(MarkDef) The overall opacity (value between [0,1]).

Default value: 0.7 for non-aggregate plots with point, tick, circle, or square marks or layered bar charts and 1 otherwise.

order
(MarkDef) For line and trail marks, this order property can be set to null or false to make the lines use the original order in the data sources.

orient
(MarkDef) The orientation of a non-stacked bar, tick, area, and line charts. The value is either horizontal (default) or vertical.

For bar, rule and tick, this determines whether the size of the bar and tick should be applied to x or y dimension.

For area, this property determines the orient property of the Vega output.

For line and trail marks, this property determines the sort order of the points in the line if config.sortLineBy is not specified. For stacked charts, this is always determined by the orientation of the stack; therefore explicitly specified value will be ignored.

point
(MarkDef) A flag for overlaying points on top of line or area marks, or an object defining the properties of the overlayed points.

If this property is "transparent", transparent points will be used (for enhancing tooltips and selections).

If this property is an empty object ({}) or true, filled points with default properties will be used.

If this property is false, no points would be automatically added to line or area marks.

Default value: false.

radius
(MarkDef) Polar coordinate radial offset, in pixels, of the text label from the origin determined by the x and y properties.

shape
(MarkDef) Shape of the point marks. Supported values include:

plotting shapes: "circle", "square", "cross", "diamond", "triangle-up", "triangle-down", "triangle-right", or "triangle-left".

the line symbol "stroke"

centered directional shapes "arrow", "wedge", or "triangle"

a custom SVG path string (For correct sizing, custom shape paths should be defined within a square bounding box with coordinates ranging from -1 to 1 along both the x and y dimensions.)

Default value: "circle"

size
(MarkDef) Default size for marks.

For point/circle/square, this represents the pixel area of the marks. For example: in the case of circles, the radius is determined in part by the square root of the size value.

For bar, this represents the band size of the bar, in pixels.

For text, this represents the font size, in pixels.

Default value: 30 for point, circle, square marks; rangeStep - 1 for bar marks with discrete dimensions; 5 for bar marks with continuous dimensions; 11 for text marks.

stroke
(MarkDef) Default Stroke Color. This has higher precedence than config.color

Default value: (None)

strokeCap
(MarkDef) The stroke cap for line ending style. One of "butt", "round", or "square".

Default value: "square"

strokeDash
(MarkDef) An array of alternating stroke, space lengths for creating dashed or dotted lines.

strokeDashOffset
(MarkDef) The offset (in pixels) into which to begin drawing with the stroke dash array.

strokeJoin
(MarkDef) The stroke line join method. One of "miter", "round" or "bevel".

Default value: "miter"

strokeMiterLimit
(MarkDef) The miter limit at which to bevel a line join.

strokeOpacity
(MarkDef) The stroke opacity (value between [0,1]).

Default value: 1

strokeWidth
(MarkDef) The stroke width, in pixels.

style
(MarkDef) A string or array of strings indicating the name of custom styles to apply to the mark. A style is a named collection of mark property defaults defined within the style configuration. If style is an array, later styles will override earlier styles. Any mark properties explicitly defined within the encoding will override a style default.

Default value: The mark's name. For example, a bar mark will have style "bar" by default. Note: Any specified style will augment the default style. For example, a bar mark with "style": "foo" will receive from config.style.bar and config.style.foo (the specified style "foo" has higher precedence).

tension
(MarkDef) Depending on the interpolation type, sets the tension parameter (for line and area marks).

text
(MarkDef) Placeholder text if the text channel is not specified

theta
(MarkDef) Polar coordinate angle, in radians, of the text label from the origin determined by the x and y properties. Values for theta follow the same convention of arc mark startAngle and endAngle properties: angles are measured in radians, with 0 indicating "north".

thickness
(MarkDef) Thickness of the tick mark.

Default value: 1

tooltip
(MarkDef) The tooltip text string to show upon mouse hover or an object defining which fields should the tooltip be derived from.

If tooltip is {"content": "encoding"}, then all fields from encoding will be used.

If tooltip is {"content": "data"}, then all fields that appear in the highlighted data point will be used.

If set to null, then no tooltip will be used.

x
(MarkDef) X coordinates of the marks, or width of horizontal "bar" and "area" without x2.

x2
(MarkDef) X2 coordinates for ranged "area", "bar", "rect", and "rule".

x2Offset
(MarkDef) Offset for x2-position.

xOffset
(MarkDef) Offset for x-position.

y
(MarkDef) Y coordinates of the marks, or height of vertical "bar" and "area" without y2

y2
(MarkDef) Y2 coordinates for ranged "area", "bar", "rect", and "rule".

y2Offset
(MarkDef) Offset for y2-position.

yOffset
(MarkDef) Offset for y-position.

Value

A modified Vega-Lite Spec