vl_sort_opacity_by_field

Add sort by another field to encoding for opacity in a vega-lite spec.

vl_sort_opacity_by_field(spec, field = NULL, op = NULL, order = NULL)

Arguments

spec
A vega-lite spec.

field
(EncodingSortField) The data field to sort by.

Default value: If unspecified, defaults to the field specified in the outer data reference.

op
(EncodingSortField) An aggregate operation to perform on the field prior to sorting (e.g., "count", "mean" and "median"). An aggregation is required when there are multiple values of the sort field for each encoded data field. The input data objects will be aggregated, grouped by the encoded data field.

For a full list of operations, please see the documentation for aggregate.

Default value: "sum" for stacked plots. Otherwise, "mean".

order
(EncodingSortField) The sort order. One of "ascending" (default), "descending", or null (no not sort).

Value

A modified spec

See also