vl_add_properties

Add properties to top level of a vega-lite spec. Allows adding properties like width,height, background which don't have a specific function for adding them (unlike markor encoding).

Usage

vl_add_properties(
  spec,
  align = NULL,
  autosize = NULL,
  background = NULL,
  bounds = NULL,
  center = NULL,
  datasets = NULL,
  description = NULL,
  height = NULL,
  name = NULL,
  padding = NULL,
  params = NULL,
  projection = NULL,
  spacing = NULL,
  title = NULL,
  usermeta = NULL,
  view = NULL,
  width = NULL,
  columns = NULL,
  .object = NULL
)

Arguments

spec

An input vega-lite spec

align

The alignment to apply to grid rows and columns. The supported string values are "all", "each", and "none".

For "none", a flow layout will be used, in which adjacent subviews are simply placed one after the other.

For "each", subviews will be aligned into a clean grid structure, but each row or column may be of variable size.

For "all", subviews will be aligned and each row or column will be sized identically based on the maximum observed size. String values for this property will be applied to both grid rows and columns.

Alternatively, an object value of the form {"row": string, "column": string} can be used to supply different alignments for rows and columns.

Default value: "all".

autosize

How the visualization size should be determined. If a string, should be one of "pad", "fit" or "none". Object values can additionally specify parameters for content sizing and automatic resizing.

Default value: pad

background

CSS color property to use as the background of the entire view.

Default value: "white"

bounds

The bounds calculation method to use for determining the extent of a sub-plot. One of full (the default) or flush.

If set to full, the entire calculated bounds (including axes, title, and legend) will be used.

If set to flush, only the specified width and height values for the sub-view will be used. The flush setting can be useful when attempting to place sub-plots without axes or legends into a uniform grid structure.

Default value: "full"

center

Boolean flag indicating if subviews should be centered relative to their respective rows or columns.

Default value: false

Or: Boolean flag indicating if subviews should be centered relative to their respective rows or columns.

An object value of the form {"row": boolean, "column": boolean} can be used to supply different centering values for rows and columns.

Default value: false

datasets

A global data store for named datasets. This is a mapping from names to inline datasets. This can be an array of objects or primitive values or a string. Arrays of primitive values are ingested as objects with a data property.

description

Description of this mark for commenting purpose.

height

The height of a visualization.

For a plot with a continuous y-field, height should be a number.

For a plot with either a discrete y-field or no y-field, height can be either a number indicating a fixed height or an object in the form of {step: number} defining the height per discrete step. (No y-field is equivalent to having one discrete step.)

To enable responsive sizing on height, it should be set to "container".

Default value: Based on config.view.continuousHeight for a plot with a continuous y-field and config.view.discreteHeight otherwise.

Note: For plots with row and column channels, this represents the height of a single view and the "container" option cannot be used.

See also: height documentation.

name

Name of the visualization for later reference.

padding

The default visualization padding, in pixels, from the edge of the visualization canvas to the data rectangle. If a number, specifies padding for all sides. If an object, the value should have the format {"left": 5, "top": 5, "right": 5, "bottom": 5} to specify padding for each side of the visualization.

Default value: 5

params

An array of parameters that may either be simple variables, or more complex selections that map user input to data queries.

Or: Dynamic variables or selections that parameterize a visualization.

projection

An object defining properties of geographic projection, which will be applied to shape path for "geoshape" marks and to latitude and "longitude" channels for other marks.

Or: An object defining properties of the geographic projection shared by underlying layers.

spacing

The spacing in pixels between sub-views of the composition operator. An object of the form {"row": number, "column": number} can be used to set different spacing values for rows and columns.

Default value: Depends on "spacing" property of the view composition configuration (20 by default)

Or: The spacing in pixels between sub-views of the concat operator.

Default value: 10

title

Title for the plot.

usermeta

Optional metadata that will be passed to Vega. This object is completely ignored by Vega and Vega-Lite and can be used for custom metadata.

view

An object defining the view background's fill and stroke.

Default value: none (transparent)

width

The width of a visualization.

For a plot with a continuous x-field, width should be a number.

For a plot with either a discrete x-field or no x-field, width can be either a number indicating a fixed width or an object in the form of {step: number} defining the width per discrete step. (No x-field is equivalent to having one discrete step.)

To enable responsive sizing on width, it should be set to "container".

Default value: Based on config.view.continuousWidth for a plot with a continuous x-field and config.view.discreteWidth otherwise.

Note: For plots with row and column channels, this represents the width of a single view and the "container" option cannot be used.

See also: width documentation.

columns

The number of columns to include in the view composition layout.

Default value: undefined -- An infinite number of columns (a single row) will be assumed. This is equivalent to hconcat (for concat) and to using the column channel (for facet and repeat).

Note:

This property is only for:

the general (wrappable) concat operator (not hconcat/vconcat)

the facet and repeat operator with one field/repetition definition (without row/column nesting)

Setting the columns to 1 is equivalent to vconcat (for concat) and to using the row channel (for facet and repeat).

.object

Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than spec. Objects can be of type: TopLevelUnitSpec, TopLevelFacetSpec, TopLevelLayerSpec, TopLevelRepeatSpec, TopLevelConcatSpec, TopLevelVConcatSpec or TopLevelHConcatSpec

Value

A modified Vega-Lite Spec