vl_add_properties

Add properties to top level of a vega-lite spec. Allows adding properties like width,height, background which don't have a specific function for adding them (unlike markor encoding).

vl_add_properties(spec, align = NULL, autosize = NULL,
  background = NULL, bounds = NULL, center = NULL, columns = NULL,
  datasets = NULL, description = NULL, height = NULL, name = NULL,
  padding = NULL, projection = NULL, spacing = NULL, title = NULL,
  usermeta = NULL, view = NULL, width = NULL)

Arguments

spec
An input vega-lite spec

align
(TopLevelUnitSpec, TopLevelFacetSpec, TopLevelRepeatSpec, TopLevelConcatSpec) The alignment to apply to grid rows and columns. The supported string values are "all", "each", and "none".

For "none", a flow layout will be used, in which adjacent subviews are simply placed one after the other.

For "each", subviews will be aligned into a clean grid structure, but each row or column may be of variable size.

For "all", subviews will be aligned and each row or column will be sized identically based on the maximum observed size. String values for this property will be applied to both grid rows and columns.

Alternatively, an object value of the form {"row": string, "column": string} can be used to supply different alignments for rows and columns.

Default value: "all".

autosize
(TopLevelUnitSpec, TopLevelFacetSpec, TopLevelLayerSpec, TopLevelRepeatSpec, TopLevelConcatSpec, TopLevelVConcatSpec, TopLevelHConcatSpec) Sets how the visualization size should be determined. If a string, should be one of "pad", "fit" or "none". Object values can additionally specify parameters for content sizing and automatic resizing. "fit" is only supported for single and layered views that don't use rangeStep.

Default value: pad

background
(TopLevelUnitSpec, TopLevelFacetSpec, TopLevelLayerSpec, TopLevelRepeatSpec, TopLevelConcatSpec, TopLevelVConcatSpec, TopLevelHConcatSpec) CSS color property to use as the background of the entire view.

Default value: none (transparent)

bounds
(TopLevelUnitSpec, TopLevelFacetSpec, TopLevelRepeatSpec, TopLevelConcatSpec, TopLevelVConcatSpec, TopLevelHConcatSpec) The bounds calculation method to use for determining the extent of a sub-plot. One of full (the default) or flush.

If set to full, the entire calculated bounds (including axes, title, and legend) will be used.

If set to flush, only the specified width and height values for the sub-view will be used. The flush setting can be useful when attempting to place sub-plots without axes or legends into a uniform grid structure.

Default value: "full"

center
(TopLevelVConcatSpec, TopLevelHConcatSpec) Boolean flag indicating if subviews should be centered relative to their respective rows or columns.

Default value: false

(TopLevelUnitSpec, TopLevelFacetSpec, TopLevelRepeatSpec, TopLevelConcatSpec) Boolean flag indicating if subviews should be centered relative to their respective rows or columns.

An object value of the form {"row": boolean, "column": boolean} can be used to supply different centering values for rows and columns.

Default value: false

columns
(TopLevelUnitSpec, TopLevelFacetSpec, TopLevelRepeatSpec, TopLevelConcatSpec) The number of columns to include in the view composition layout.

Default value: undefined -- An infinite number of columns (a single row) will be assumed. This is equivalent to hconcat (for concat) and to using the column channel (for facet and repeat).

Note:

This property is only for:

the general (wrappable) concat operator (not hconcat/vconcat)

the facet and repeat operator with one field/repetition definition (without row/column nesting)

Setting the columns to 1 is equivalent to vconcat (for concat) and to using the row channel (for facet and repeat).

datasets
(TopLevelUnitSpec, TopLevelFacetSpec, TopLevelLayerSpec, TopLevelRepeatSpec, TopLevelConcatSpec, TopLevelVConcatSpec, TopLevelHConcatSpec) A global data store for named datasets. This is a mapping from names to inline datasets. This can be an array of objects or primitive values or a string. Arrays of primitive values are ingested as objects with a data property.

description
(TopLevelUnitSpec, TopLevelFacetSpec, TopLevelLayerSpec, TopLevelRepeatSpec, TopLevelConcatSpec, TopLevelVConcatSpec, TopLevelHConcatSpec) Description of this mark for commenting purpose.

height
(TopLevelUnitSpec, TopLevelLayerSpec) The height of a visualization.

Default value:

If a view's autosize type is "fit" or its y-channel has a continuous scale, the height will be the value of config.view.height.

For y-axis with a band or point scale: if rangeStep is a numeric value or unspecified, the height is determined by the range step, paddings, and the cardinality of the field mapped to y-channel. Otherwise, if the rangeStep is null, the height will be the value of config.view.height.

If no field is mapped to y channel, the height will be the value of rangeStep.

Note: For plots with row and column channels, this represents the height of a single view.

See also: The documentation for width and height contains more examples.

name
(TopLevelUnitSpec, TopLevelFacetSpec, TopLevelLayerSpec, TopLevelRepeatSpec, TopLevelConcatSpec, TopLevelVConcatSpec, TopLevelHConcatSpec) Name of the visualization for later reference.

padding
(TopLevelUnitSpec, TopLevelFacetSpec, TopLevelLayerSpec, TopLevelRepeatSpec, TopLevelConcatSpec, TopLevelVConcatSpec, TopLevelHConcatSpec) The default visualization padding, in pixels, from the edge of the visualization canvas to the data rectangle. If a number, specifies padding for all sides. If an object, the value should have the format {"left": 5, "top": 5, "right": 5, "bottom": 5} to specify padding for each side of the visualization.

Default value: 5

projection
(TopLevelUnitSpec) An object defining properties of geographic projection, which will be applied to shape path for "geoshape" marks and to latitude and "longitude" channels for other marks.

(TopLevelLayerSpec) An object defining properties of the geographic projection shared by underlying layers.

spacing
(TopLevelUnitSpec, TopLevelFacetSpec, TopLevelRepeatSpec, TopLevelConcatSpec) The spacing in pixels between sub-views of the composition operator. An object of the form {"row": number, "column": number} can be used to set different spacing values for rows and columns.

Default value: Depends on "spacing" property of the view composition configuration (20 by default)

(TopLevelVConcatSpec, TopLevelHConcatSpec) The spacing in pixels between sub-views of the concat operator.

Default value: 10

title
(TopLevelUnitSpec, TopLevelFacetSpec, TopLevelLayerSpec, TopLevelRepeatSpec, TopLevelConcatSpec, TopLevelVConcatSpec, TopLevelHConcatSpec) Title for the plot.

usermeta
(TopLevelUnitSpec, TopLevelFacetSpec, TopLevelLayerSpec, TopLevelRepeatSpec, TopLevelConcatSpec, TopLevelVConcatSpec, TopLevelHConcatSpec) Optional metadata that will be passed to Vega. This object is completely ignored by Vega and Vega-Lite and can be used for custom metadata.

view
(TopLevelUnitSpec, TopLevelLayerSpec) An object defining the view background's fill and stroke.

Default value: none (transparent)

width
(TopLevelUnitSpec, TopLevelLayerSpec) The width of a visualization.

Default value: This will be determined by the following rules:

If a view's autosize type is "fit" or its x-channel has a continuous scale, the width will be the value of config.view.width.

For x-axis with a band or point scale: if rangeStep is a numeric value or unspecified, the width is determined by the range step, paddings, and the cardinality of the field mapped to x-channel. Otherwise, if the rangeStep is null, the width will be the value of config.view.width.

If no field is mapped to x channel, the width will be the value of config.scale.textXRangeStep for text mark and the value of rangeStep for other marks.

Note: For plots with row and column channels, this represents the width of a single view.

See also: The documentation for width and height contains more examples.

Value

A modified Vega-Lite Spec