vl_config_title

Add title config (TitleConfig) to a vega-lite spec.

vl_config_title(spec, .object = NULL, align = NULL, anchor = NULL,
  angle = NULL, baseline = NULL, color = NULL, dx = NULL,
  dy = NULL, font = NULL, fontSize = NULL, fontStyle = NULL,
  fontWeight = NULL, frame = NULL, limit = NULL, lineHeight = NULL,
  offset = NULL, orient = NULL, subtitleColor = NULL,
  subtitleFont = NULL, subtitleFontSize = NULL,
  subtitleFontStyle = NULL, subtitleFontWeight = NULL,
  subtitleLineHeight = NULL, subtitlePadding = NULL)

Arguments

spec
An input vega-lite spec

.object
Directly input an object, rather than creating one via the other arguments. Should not be used in conjunction with the other arguments other than 'spec'. Objects can be of type: ExcludeMappedValueRef<BaseTitle>

align
(ExcludeMappedValueRef<BaseTitle>) Horizontal text alignment for title text. One of "left", "center", or "right".

anchor
(ExcludeMappedValueRef<BaseTitle>) The anchor position for placing the title and subtitle text. One of "start", "middle", or "end". For example, with an orientation of top these anchor positions map to a left-, center-, or right-aligned title.

angle
(ExcludeMappedValueRef<BaseTitle>) Angle in degrees of title and subtitle text.

baseline
(ExcludeMappedValueRef<BaseTitle>) Vertical text baseline for title and subtitle text. One of "top", "middle", "bottom", or "alphabetic".

color
(ExcludeMappedValueRef<BaseTitle>) Text color for title text.

dx
(ExcludeMappedValueRef<BaseTitle>) Delta offset for title and subtitle text x-coordinate.

dy
(ExcludeMappedValueRef<BaseTitle>) Delta offset for title and subtitle text y-coordinate.

font
(ExcludeMappedValueRef<BaseTitle>) Font name for title text.

fontSize
(ExcludeMappedValueRef<BaseTitle>) Font size in pixels for title text.

fontStyle
(ExcludeMappedValueRef<BaseTitle>) Font style for title text.

fontWeight
(ExcludeMappedValueRef<BaseTitle>) Font weight for title text. This can be either a string (e.g "bold", "normal") or a number (100, 200, 300, ..., 900 where "normal" = 400 and "bold" = 700).

frame
(ExcludeMappedValueRef<BaseTitle>) The reference frame for the anchor position, one of "bounds" (to anchor relative to the full bounding box) or "group" (to anchor relative to the group width or height).

limit
(ExcludeMappedValueRef<BaseTitle>) The maximum allowed length in pixels of title and subtitle text.

lineHeight
(ExcludeMappedValueRef<BaseTitle>) Line height in pixels for multi-line title text.

offset
(ExcludeMappedValueRef<BaseTitle>) The orthogonal offset in pixels by which to displace the title group from its position along the edge of the chart.

orient
(ExcludeMappedValueRef<BaseTitle>) Default title orientation ("top", "bottom", "left", or "right")

subtitleColor
(ExcludeMappedValueRef<BaseTitle>) Text color for subtitle text.

subtitleFont
(ExcludeMappedValueRef<BaseTitle>) Font name for subtitle text.

subtitleFontSize
(ExcludeMappedValueRef<BaseTitle>) Font size in pixels for subtitle text.

subtitleFontStyle
(ExcludeMappedValueRef<BaseTitle>) Font style for subtitle text.

subtitleFontWeight
(ExcludeMappedValueRef<BaseTitle>) Font weight for subtitle text. This can be either a string (e.g "bold", "normal") or a number (100, 200, 300, ..., 900 where "normal" = 400 and "bold" = 700).

subtitleLineHeight
(ExcludeMappedValueRef<BaseTitle>) Line height in pixels for multi-line subtitle text.

subtitlePadding
(ExcludeMappedValueRef<BaseTitle>) The padding in pixels between title and subtitle text.

Value

A modified Vega-Lite Spec